% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaLike.R
\name{anovaLike}
\alias{anovaLike}
\title{A function allowing the identification of differentially expressed genes.}
\usage{
anovaLike(
  group = c("sudo", "docker"),
  file,
  sep,
  cluster.file,
  ref.cluster,
  logFC.threshold = 1,
  FDR.threshold,
  logCPM.threshold = 4,
  plot = c(TRUE, FALSE)
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{file, }{a character string indicating the counts table file with the path of the file.}

\item{sep, }{separator used in count file, e.g. '\\t', ','}

\item{cluster.file, }{a character string indicating the _clustering.output.txt file of interest, generated by bootstrapSimlar or bootStrapTsne. IMPORTANT this file miust be located in the same folder where counts.table is placed}

\item{ref.cluster, }{a number indicating the cluster to be used a reference for anova-like comparison with the other clusters.}

\item{logFC.threshold, }{minimal logFC present in at least one of the comparisons with respect to reference covariate}

\item{FDR.threshold, }{minimal FDR present in at least one of the comparisons with respect to reference covariate}

\item{logCPM.threshold, }{minimal average abundance}

\item{plot, }{boolean, TRUE a plot of differentially expressed genes is generated}
}
\value{
Three tab delimited files file with prefix DE\_, filtered\_DE\_, logFC\_filtered\_DE\_ followed by the counts table name, the count table, reordered on the basis of cluster positions, has the extension \_reordered.txt
}
\description{
This function executes in a docker edgeR for the idnetification of differentially expressed genes in single-cells RNAseq
}
\examples{
\dontrun{
    #running deDetection
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5_clustering.output.txt")
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")
anovaLike(group="docker", file=paste(getwd(),"annotated_setPace_10000_noC5.txt",sep="/"),
       sep="\t", cluster.file="annotated_setPace_10000_noC5_clustering.output.txt", ref.cluster=3,
       logFC.threshold=1, FDR.threshold=0.05, logCPM.threshold=4, plot=TRUE)

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
