% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocluster4clustering.R
\name{autocluster4clustering}
\alias{autocluster4clustering}
\title{Autoencoder clustering}
\usage{
autocluster4clustering(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  projectName,
  clusterMethod = c("GRIPH", "SIMLR", "SEURAT", "SHARP", "KMEANS", "HCLUST", "FLOW"),
  seed = 1111,
  pcaDimensions = 20,
  perplexity = 20
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included.Has to be the one in the projectName folder.Different so from the previous one.}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{number of cluster in which the dataset is divided}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{clusterMethod, }{clustering methods: "GRIPH","SIMLR","SEURAT","SHARP","KMEANS","HCLUST","FLOW"}

\item{seed, }{important value to reproduce the same results with same input}

\item{pcaDimensions, }{number of dimensions to use for Seurat Pca reduction.}

\item{perplexity, }{value for cluster aggregation, required for KMEANS}
}
\value{

}
\description{
This function Compress data using autoencoder partially connected
}
\examples{
\dontrun{
 autocluster4clustering(group=c("sudo"), scratch.folder=scratch, file=file, separator=",", nCluster=3, projectName=projectName, clusterMethod=i,seed=1111,pcaDimensions=20)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
