% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder4Pseudobulk.R
\name{autoencoder4pseudoBulk}
\alias{autoencoder4pseudoBulk}
\title{Autoencoder for pseudoBulk}
\usage{
autoencoder4pseudoBulk(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  permutation,
  nEpochs,
  patiencePercentage = 5,
  seed = 1111,
  projectName,
  bN,
  lr = 0.01,
  beta_1 = 0.9,
  beta_2 = 0.999,
  epsilon = 1e-08,
  decay = 0,
  loss = "mean_squared_error",
  regularization = 10
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{permutation, }{number of permutations to perform the pValue to evaluate clustering}

\item{nEpochs, }{number of Epochs for neural network training}

\item{patiencePercentage, }{number of Epochs percentage of not training before to stop.}

\item{seed, }{important value to reproduce the same results with same input}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{bN, }{path to the clustering.output file}

\item{lr, }{learning rate, the speed of learning. Higher value may increase the speed of convergence but may also be not very precise}

\item{beta_1, }{look at keras optimizer parameters}

\item{beta_2, }{look at keras optimizer parameters}

\item{epsilon, }{look at keras optimizer parameters}

\item{decay, }{look at keras optimizer parameters}

\item{loss, }{loss of function to use, for other loss of function check the keras loss of functions.}

\item{regularization, }{this parameter balances between reconstruction loss and enforcing a normal distribution in the latent space.}
}
\value{

}
\description{
The present function compress data using autoencoder partially connected creating pseudoBulk matrix
}
\examples{
\dontrun{
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
