% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoClusters.R
\name{autoencoderClustering}
\alias{autoencoderClustering}
\title{Autoencoder clustering}
\usage{
autoencoderClustering(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  projectName,
  clusterMethod = c("GRIPH", "SIMLR", "SEURAT", "SHARP"),
  seed = 1111,
  pcaDimensions,
  permAtTime = 4,
  largeScale = FALSE
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included.Has to be the one in the projectName folder.Different so from the previous one.}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{number of cluster in which the dataset is divided}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{clusterMethod, }{clustering methods: "GRIPH","SIMLR","SEURAT","SHARP"}

\item{seed, }{important value to reproduce the same results with same input}

\item{pcaDimensions, }{number of dimensions to use for Seurat Pca reduction.}

\item{permAtTime, }{number of permutation in parallel}

\item{largeScale, }{boolean for SIMLR analysis, TRUE if rows are less then columns or if the computational time are huge}
}
\description{
This function Compress data using autoencoder partially connected
}
\examples{
\dontrun{
 autoencoderClustering(group="docker", scratch.folder="/home/user/Riccardo/Riccardo/1_inDocker_2/scratch", file="/home/user/Riccardo/Riccardo/1_inDocker_2/data/Results/testDocker/setA.csv",separator=",", nCluster=5,clusterMethod=c("SEURAT"),seed=1111,projectName="testDocker",13, largeScale = FALSE)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
