% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoderDB.R
\name{autoencoderDB}
\alias{autoencoderDB}
\title{Autoencoder4clustering}
\usage{
autoencoderDB(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  permutation,
  nEpochs,
  patiencePercentage = 5,
  seed = 1111,
  projectName,
  lr = 0.01,
  loss = "mean_squared_error"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{permutation, }{number of permutations to perform the pValue to evaluate clustering}

\item{nEpochs, }{number of Epochs for neural network training}

\item{patiencePercentage, }{number of Epochs percentage of not training before to stop.}

\item{seed, }{important value to reproduce the same results with same input}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{lr, }{learning rate, the speed of learning. Higher value may increase the speed of convergence but may also be not very precise}

\item{loss, }{loss of function to use, for other loss of function check the keras loss of functions.}
}
\value{

}
\description{
The present function compress data using autoencoder partially connected
}
\examples{
\dontrun{
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
