% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFeature.R
\name{autoFeature}
\alias{autoFeature}
\title{Feature selection for autoencoder}
\usage{
autoFeature(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  projectName
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included.Has to be the one in the projectName folder.Different so from the previous one.}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{number of cluster in which the dataset is divided}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}
}
\description{
This function extract feature selection for each clusters.
}
\examples{
\dontrun{
 autoFeature(group="docker", scratch.folder="/home/user/Riccardo/Riccardo/1_inDocker_2/scratch", file="/home/user/Riccardo/Riccardo/1_inDocker_2/data/Results/testDocker/setA.csv",separator=",", nCluster=5, projectName="testDocker")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
