% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeSpacePermutation.R
\name{bayeSpacePermutation}
\alias{bayeSpacePermutation}
\title{bayeSpace Permutation}
\usage{
bayeSpacePermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  filtered_feature_bc_matrix,
  n_clusters,
  pcaDimensions = 10,
  spatial,
  nPerm = 80,
  permAtTime = 8,
  percent = 10,
  seed = 1111
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the full path + name with extension (.txt) of the expression matrix}

\item{filtered_feature_bc_matrix, }{a character string indicating the full path of the Space Ranger output directory (10X published data)}

\item{n_clusters, }{an integer indicating how many clusters BayeSpace should search}

\item{pcaDimensions, }{number of PCA dimension to keep}

\item{spatial, }{a character string indicating the full path of the directory containing tissue images and positional data (10X published data)}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{seed, }{important value to reproduce the same results with same input}
}
\value{
To write
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering 
 using bayeSpace
}
\author{
Giovanni Motterle, giovanni [dot] motterle [at] studenti [dot] univr [dot] it, University of Verona
}
