% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCsc_function.R
\name{BCscTool}
\alias{BCscTool}
\title{BCscTool}
\usage{
BCscTool(
  group = c("sudo", "docker"),
  scratch.folder,
  setA,
  setAclustering,
  markerA,
  setB,
  setBclustering,
  markerB,
  separator,
  contamination = 5,
  permutation = 50,
  threshold = 3
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{setA, }{path to the first dataset}

\item{setAclustering, }{path to the clustering file for the first dataset. This file is derived from clustering.output file generated by the rCASC clustering. It contains the following columns: CellName, Belonging_Cluster}

\item{markerA, }{path to the file containing cluster specific genes for the first dataset. The file requires the following column: gene, cluster}

\item{setB, }{path to the second dataset}

\item{setBclustering, }{path to the clustering file for the second dataset. This file is derived from clustering.output file generated by the rCASC clustering. It contains the following columns: CellName, Belonging_Cluster}

\item{markerB, }{path to the file containing cluster specific genes for the second dataset. The file requires the following column: gene, cluster}

\item{separator, }{separator used in both count file, e.g. ','}

\item{contamination, }{fraction of genes to be contaminated at each eprmutaitone DEFAULT VALUE: 5}

\item{permutation, }{number of permutations DEFAULT VALUE: 50}

\item{threshold, }{minimum value to consider a gene expressed in both count sets. DEFAULT VALUE: 3}
}
\value{

}
\description{
The present function associates the clusters of two different and indipendent experiments using the Bray Curtis dissimilarity. BOTH set must have same number and name of ROW (GENES)
}
\examples{
		
\dontrun{
BCscTool(group = "docker",
         scratch.folder = "/home/user09/piacenti/ProvaBCSC/scratch",
         setA = paste(getwd(),"setA.csv",sep = "/"),
         setAclustering = paste(getwd(),"setA_clustering.csv",sep = "/"),
         markerA = paste(getwd(),"setA.markers.csv",sep = "/"),
         setB =paste(getwd(),"set1.csv",sep = "/"),
         setBclustering = paste(getwd(),"set1_clustering.csv",sep = "/"),
         markerB =paste(getwd(),"set1.markersPROVA.csv",sep = "/"),
         separator =",",
         contamination = 10 ,
         permutation = 10,
         threshold = 3 )
}
}
\author{
Gabriele Piacenti, g [dot] pia91 [at] gmail [dot] com, University of Torino
}
