% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCsc_wrapper.R
\name{BCscWrapper}
\alias{BCscWrapper}
\title{BCscTool_wrapper}
\usage{
BCscWrapper(
  group = c("sudo", "docker"),
  setX,
  nclustX,
  markerX,
  setY,
  nclustY,
  markerY,
  separator = ",",
  scratch.folder,
  output.folder,
  pvalueComet = 1,
  subdivision = 10,
  starting_genes = 500,
  contamination = 5,
  permutation = 50,
  threshold = 3
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{setX, }{path to the first dataset}

\item{nclustX, }{number of cluster of the first dataset.}

\item{markerX, }{path to the FOLDER containing cluster-specific genes for the FIRST dataset. (after using Comet directory '.../nclustX/outputdata')}

\item{setY, }{path to the second dataset}

\item{nclustY, }{number of cluster of the second dataset}

\item{markerY, }{path to the FOLDER containing cluster-specific genes for the SECOND dataset. (after using Comet directory '.../nclustY/outputdata')}

\item{separator, }{separator used in both count file, e.g. '\\t', ','}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{output.folder, }{a character string indicating the path of the output folder to save analysis}

\item{pvalueComet, }{parameter to select only marker genes  with a pvalue under a threshold value. DEFAULT = 1}

\item{subdivision, }{number of progressive reductions of the starting genes list. DEFAULT = 10}

\item{starting_genes, }{number of starting genes for each cluster selected from the main marker gene lists. DEFAULT = 500}

\item{contamination, }{percentage of genes to be contaminated at each permutation during the Bray-Curtis calculation. DEFAULT = 5}

\item{permutation, }{number of Bray-Curtis permutations. DEFAULT = 50}

\item{threshold, }{minimum value to consider a gene expressed in both count sets. DEFAULT VALUE: 3}
}
\value{

}
\description{
The present function associates the clusters of two different and indipendent experiments using the Bray Curtis dissimilarity. Starting from the two original sets it takes only the genes common to both and calculates the Bray Curtis Dissimilarity between the clusters of one set and those of the other. all the results will be saved in the folder called "BCsc" inside the folder where the files of the first set are located
}
\examples{
		
\dontrun{
BCscWrapper(group="docker",
      setX = '/2tb/torino/piacenti/NT1_NT2/CRC0327_NT_1_clx/VandE/VandE.csv',
      nclustX = 7,
      markerX = '/2tb/torino/piacenti/NT1_NT2/CRC0327_NT_1_clx/VandE/Results/VandE/7/outputdata',
      setY = '/2tb/torino/piacenti/NT1_NT2/CRC0327_NT_2_clx/VandE/VandE.csv',
      nclustY = 8,
     markerY = '/2tb/torino/piacenti/NT1_NT2/CRC0327_NT_2_clx/VandE/Results/VandE/8/outputdata',
     scratch.folder = '/2tb/torino/piacenti/user09/piacenti/Prova_Prove/Script/scratch',
     output.folder = '/2tb/torino/piacenti/user09/piacenti/Prova_Prove/Script',
     pvalueComet = 1,
     subdivision = 10,
     starting_genes = 500,
     contamination = 5,
     permutation = 50,
     threshold = 3)
}
}
\author{
Gabriele Piacenti, g [dot] pia91 [at] gmail [dot] com, University of Torino
}
