% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapsVideo.R
\name{bootstrapsVideo}
\alias{bootstrapsVideo}
\title{Bootstraps vide}
\usage{
bootstrapsVideo(group = c("sudo", "docker"), scratch.folder, file,
  nCluster, separator, framePP = 200, permutationNumber)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{which nCluster results to use for this analysis}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{framePP, }{Number of frames for each bootstrap}

\item{permutationNumber, }{Number of random bootstraps, should be less or equal to the total number of permutations used to generate tha clustering data}
}
\value{
a video
}
\description{
This function generate a video showing the relocation of cells during the bootstraps
}
\examples{
\dontrun{
 bootstrapsVideo(group,scratch.folder,file,nCluster,separator)#
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
