% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkClusters.R
\name{bulkClusters}
\alias{bulkClusters}
\title{bulkClusters}
\usage{
bulkClusters(group = c("sudo", "docker"), scratch.folder, file, separator, cl)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{matrix separator, ',', '\\t'}

\item{cl, }{name and path of the file clustering.output previously generated from clustering algorithm from rcasc}
}
\description{
The present function create pseudo-bulk matrix from clustering.output file. The output are three files: _bulklog2, which is not normalized, _bulkColumn, which is z-scoere calculated over each column, _bulkRow, which is z-score calculated over each row
}
\examples{
\dontrun{
 bulkClusters(group="docker", scratch.folder="/home/user/scratch", 
 file="/home/user/temp/setA.csv",separator=",",
 cl="/home/user/temp/Results/setA/3/setA_clustering.output.csv")
 
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
