% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_count.R
\name{cellranger_count}
\alias{cellranger_count}
\title{Cellranger count}
\usage{
cellranger_count(group = c("sudo", "docker"), id, transcriptome.folder,
  fastq.folder, sample, expect.cells = 3000, nosecondary = FALSE,
  indices = NULL, scratch.folder)
}
\arguments{
\item{id}{A unique run ID string: e.g. sample345}

\item{sample}{Sample name as specified in the sample sheet supplied to cellranger mkfastq. Allowable characters in sample names are letters, numbers, hyphens, and underscores.}

\item{nosecondary}{(optional) Boulean TRUE/FAlSE. Add this flag to skip secondary analysis of the gene-barcode matrix (dimensionality reduction, clustering and visualization). Set this if you plan to use cellranger reanalyze or your own custom analysis.}

\item{scratch.folder}{scratch folder path}

\item{fastq}{Path of the fastq_path folder}

\item{transcriptome}{Path to the Cell Ranger compatible transcriptome reference downloadable from 10xGenomics repository, for more info see CASC vignette}

\item{expect-cells}{(optional) Expected number of recovered cells. Default: 3,000 cells.}
}
\description{
This function takes FASTQ files from cellranger mkfastq and performs alignment, filtering, barcode counting, and UMI counting. It uses the Chromium cellular barcodes to generate gene-barcode matrices, determine clusters, and perform gene expression analysis. The count pipeline can take input from multiple sequencing runs on the same library.
}
\examples{
\dontrun{
home <- getwd()
library(casc)
downloadContainers()
setwd("/data/genomes/cellranger_hg19mm10")
#getting the human and mouse cellranger index
system("wget http://cf.10xgenomics.com/supp/cell-exp/refdata-cellranger-hg19-and-mm10-2.1.0.tar.gz")
setwd(home)
# 100 cells 1:1 Mixture of Fresh Frozen Human (HEK293T) and Mouse (NIH3T3) Cells
system("wget http://cf.10xgenomics.com/samples/cell-exp/2.1.0/hgmm_100/hgmm_100_fastqs.tar")
cellranger_count(group="docker",  id="hgmm", transcriptome.folder="/data/genomes/cellranger_hg19mm10",  fastq.folder="/data/test_cell_ranger/fastqs",  sample="hgmm", expect.cells=100, nosecondary=TRUE, scratch.folder="/data/scratch")




}


}
\author{
Greta Romano,  romano [dot] greta [at] gmail [dot] com, University of Torino
}
