% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellrangerIndexing.R
\name{cellrangerIndexing}
\alias{cellrangerIndexing}
\title{Cellranger indexing}
\usage{
cellrangerIndexing(group = c("sudo", "docker"), scratch.folder,
  genomeFolder, gtf.url, fasta.url, bio.type = c("protein_coding",
  "unitary_pseudogene", "unprocessed_pseudogene", "processed_pseudogene",
  "transcribed_unprocessed_pseudogene", "processed_transcript",
  "antisense", "transcribed_unitary_pseudogene", "polymorphic_pseudogene",
  "lincRNA", "sense_intronic", "transcribed_processed_pseudogene",
  "sense_overlapping", "IG_V_pseudogene", "pseudogene", "TR_V_gene",
  "3prime_overlapping_ncRNA", "IG_V_gene", "bidirectional_promoter_lncRNA",
  "snRNA", "miRNA", "misc_RNA", "snoRNA", "rRNA", "IG_C_gene", "IG_J_gene",
       "TR_J_gene", "TR_C_gene", "TR_V_pseudogene", "TR_J_pseudogene",
  "IG_D_gene", "ribozyme", "IG_C_pseudogene", "TR_D_gene", "TEC",
  "IG_J_pseudogene", "scRNA", "scaRNA", "vaultRNA", "sRNA", "macro_lncRNA",
  "non_coding", "IG_pseudogene"), nThreads)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{genomeFolder, }{path for the genome folder}

\item{gtf.url, }{url for gtf download}

\item{fasta.url, }{url for fasta download}

\item{bio.type, }{biotype to filter the gtf}
}
\value{
a folder called results_cellranger, more info on the structure of this folder at https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/overview . In /somewhewre_in_your_computer/results_cellranger/outs/filtered_gene_bc_matrices the cells counts matrices results_cellranger.cvs and results_cellranger.txt are saved for further use.
}
\description{
This function creates the indexing for 10Xgenomics
}
\examples{
\dontrun{
home <- getwd()
library(rCASC)
setwd("/data/genomes/cellranger_hg19mm10")
#getting the human and mouse cellranger index
system("wget http://cf.10xgenomics.com/supp/cell-exp/refdata-cellranger-hg19-and-mm10-2.1.0.tar.gz")
untar("refdata-cellranger-hg19-and-mm10-2.1.0.tar.gz") 
setwd(home)
# 100 cells 1:1 Mixture of Fresh Frozen Human (HEK293T) and Mouse (NIH3T3) Cells
system("wget http://cf.10xgenomics.com/samples/cell-exp/2.1.0/hgmm_100/hgmm_100_fastqs.tar")
untar("hgmm_100_fastqs.tar")
home=paste(home,"/fastqs",sep="")
cellrangerCount(group="docker",  transcriptome.folder="/data/genomes/cellranger_hg19mm10/refdata-cellranger-hg19_and_mm10-2.1.0",  fastq.folder=getwd(),  expect.cells=100, nosecondary=TRUE, scratch.folder="/data/scratch", version="2")

sraDownload(group = "docker", sra.name = "SRR7762358", data.folder = getwd(), scratch.folder = "/data/scratch", threads = 8)
system("mv ./SRR7762358/SRR7762358.fastq.gz ./SRR7762358/SRR7762358_S1_L001_R1_001.fastq.gz")
cellrangerCount(group="docker",  transcriptome.folder="/data/genomes/refdata-cellranger-GRCh38-3.0.0",  fastq.folder=getwd(), sample="SRR7762358",  nosecondary=TRUE, scratch.folder="/data/scratch", version="3")

}


}
\author{
Luca Alessandrì
}
