% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringWA_Wrapper.R
\name{clusteringWA_Wrapper}
\alias{clusteringWA_Wrapper}
\title{wrapperAutoencoder}
\usage{
clusteringWA_Wrapper(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  bias = c("mirna", "TF", "CUSTOM", "kinasi", "immunoSignature", "cytoBands", "ALL"),
  permutation,
  nEpochs,
  patiencePercentage = 5,
  seed = 1111,
  projectName,
  bN = "NULL",
  lr = 0.01,
  beta_1 = 0.9,
  beta_2 = 0.999,
  epsilon = 1e-08,
  decay = 0,
  loss = "mean_squared_error",
  clusterMethod = c("GRIPH", "SIMLR", "SEURAT", "SHARP"),
  pcaDimensions = 5,
  Sp = 0.8,
  version = 2
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{number of cluster in which the dataset is divided}

\item{bias, }{bias method to use : "mirna" , "TF", "CUSTOM", kinasi,immunoSignature, cytoBands,ALL}

\item{permutation, }{number of permutations to perform the pValue to evaluate clustering}

\item{nEpochs, }{number of Epochs for neural network training}

\item{patiencePercentage, }{number of Epochs percentage of not training before to stop.}

\item{seed, }{important value to reproduce the same results with same input}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{bN, }{name of the custom bias file. This file need header, in the first column has to be the source and in the second column the gene symbol.All path needs to be provided,}

\item{lr, }{learning rate, the speed of learning. Higher value may increase the speed of convergence but may also be not very precise}

\item{beta_1, }{look at keras optimizer parameters}

\item{beta_2, }{look at keras optimizer parameters}

\item{epsilon, }{look at keras optimizer parameters}

\item{decay, }{look at keras optimizer parameters}

\item{loss, }{loss of function to use, for other loss of function check the keras loss of functions.}

\item{clusterMethod, }{clustering methods: "GRIPH","SIMLR","SEURAT","SHARP"}

\item{pcaDimensions, }{number of dimensions to use for Seurat Pca reduction.}

\item{Sp, }{minimun number of percentage of cells that has to be in common between two permutation to be the same cluster.}

\item{version, }{version 1 implements static batchsize, version 2 implements adaptive batchsize}
}
\value{
folders the complete autoencoder analysis.


}
\description{
This function executes the whole autoencoder pipeline
}
\examples{
\dontrun{
 clusteringWA_Wrapper(group=c("sudo"),scratch.folder="/home/user/autoencoderClustering/test/Scratch/",file="/home/user/autoencoderClustering/test/Data/setA.csv",separator=",",nCluster=3,bias=c("ALL"),permutation=10,nEpochs=10,patiencePercentage=5,seed=1111,projectName="TEST2",bN="NULL",lr=0.01,beta_1=0.9,beta_2=0.999,epsilon=0.00000001,decay=0.0,loss="mean_squared_error",clusterMethod=c( "GRIPH"),pcaDimensions=5,Sp=0.8, version=2)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
