% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_WADB_wrapper.R
\name{clusteringWADB_Wrapper}
\alias{clusteringWADB_Wrapper}
\title{wrapperAutoencoder}
\usage{
clusteringWADB_Wrapper(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  permutation,
  nEpochs,
  patiencePercentage = 5,
  seed = 1111,
  projectName,
  lr = 0.01,
  loss = "mean_squared_error",
  clusterMethod = c("GRIPH", "SIMLR", "SEURAT", "SHARP"),
  pcaDimensions = 5,
  Sp = 0.8
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{number of cluster in which the dataset is divided}

\item{permutation, }{number of permutations to perform the pValue to evaluate clustering}

\item{nEpochs, }{number of Epochs for neural network training}

\item{patiencePercentage, }{number of Epochs percentage of not training before to stop.}

\item{seed, }{important value to reproduce the same results with same input}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{lr, }{learning rate, the speed of learning. Higher value may increase the speed of convergence but may also be not very precise}

\item{loss, }{loss of function to use, for other loss of function check the keras loss of functions.}

\item{clusterMethod, }{clustering methods: "GRIPH","SIMLR","SEURAT","SHARP"}

\item{pcaDimensions, }{number of dimensions to use for Seurat Pca reduction.}

\item{Sp, }{minimun number of percentage of cells that has to be in common between two permutation to be the same cluster.}
}
\value{
folders the complete autoencoder analysis.
}
\description{
This function executes the whole autoencoder pipeline
}
\examples{
\dontrun{
 clusteringWADB_Wrapper(group=c("sudo"),scratch.folder="/home/user/scratch",file="/home/user/autoencoderClustering_v4/u/setA.csv",separator=",",nCluster=5,permutation=80,nEpochs=1000,patiencePercentage=5,seed=1111,projectName="yuppy",clusterMethod=c( "SIMLR"),lr=0.001)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
