% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterNgriph.R
\name{clusterNgriph}
\alias{clusterNgriph}
\title{Permutations and Clustering}
\usage{
clusterNgriph(group = c("sudo", "docker"), scratch.folder, file, nPerm,
  permAtTime, percent, separator, logTen = 0, seed = 111)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to perform}

\item{permAtTime, }{number of permutations that can be computed in parallel}

\item{percent, }{percentage of cells randomly removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{seed, }{important value to reproduce the same results with same input}
}
\value{
VioPlot of silhouette cells value for each number of cluster used,clusterP file with clustering results for each permutation, killedCell file with removed cells in each permutation, clustering.output a sommarize file with general information for each cells.
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate the range of optimal number of clusters using griph algorithm. For more info see CSC vignette.
}
\examples{
\dontrun{
 clusterNgriph(group="docker",scratch.folder="/data/scratch/",file=paste(getwd(), "bmsnkn_5x100cells.txt", sep="/"), nPerm=16, permAtTime=8, percent=10, separator="\\t",logTen=0, seed=111)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
