% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterReorg.R
\name{clusterReorg}
\alias{clusterReorg}
\title{Reorganize Cluster}
\usage{
clusterReorg(group = c("sudo", "docker"), scratch.folder, file, nCluster,
  A, B, separator, sp = 0.8)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{number of Cluster used in Kmeans to generate the clusters that you want to merge}

\item{A, }{first Cluster that has to be merged}

\item{B, }{second Cluster that has to be merged}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{sp, }{minimun number of percentage of cells that has to be in common between two permutation to be the same cluster.}
}
\value{
will change all the files generated from permAnalysis algorithm in a new folder matrixName_Cluster_merged/
}
\description{
This function executes a ubuntu docker that merge two clusters
}
\examples{
\dontrun{
clusterReorg("docker","/home/lucastormreig/CASC2.0/2.1_clusterReorg/scratch/","/home/lucastormreig/CASC2.0/2.1_clusterReorg/Data/TOTAL.csv",3,1,3,",")#
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
