% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStability.R
\name{clusterStability}
\alias{clusterStability}
\title{Permutations and Clustering}
\usage{
clusterStability(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nPerm,
  range1 = 3,
  range2 = 3,
  separator,
  logTen = 0,
  clustering,
  perplexity = 10,
  pcaDimensions,
  seed = 1111
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{range1, }{first number of cluster for k means algorithm}

\item{range2, }{last number of cluster for k means algorithm}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{clustering, }{clustering method to use : "SIMLR" , "tSne", "griph"}

\item{perplexity, }{Number of close neighbors for each point. This parameter is specific for tSne. Default value is 10.Setting this parameter when use a clustering method different by tSne will be ignored.}

\item{pcaDimensions, }{dimensions to use for pca reduction for Seurat}

\item{seed, }{important value to reproduce the same results with same input}
}
\value{
plot with cluster stability
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering.
}
\examples{
\dontrun{
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
