% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cometsc.R
\name{cometsc}
\alias{cometsc}
\title{Marker genes discovery with COMETSC}
\usage{
cometsc(group = c("sudo", "docker"), file, scratch.folder, threads = 1,
  X = 0.15, K = 2, counts = c("True", "False"), skipvis = c("True",
  "False"), nCluster, separator)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{scratch.folder, }{temporary folder where calculation is made}

\item{threads, }{integer refering to the max number of process run in parallel default 1 max the number of clusters under analysis, i.e. nCluster}

\item{X, }{from 0 to 1 argument for XL-mHG default 0.15, for more info see cometsc help.}

\item{K, }{the number of gene combinations to be considered., possible values 2, 3, 4, default 2. WARNING increasing the number of combinations makes the matrices very big}

\item{counts, }{If set to True it will graph the log(expression+1). To be used if unlogged data are provided}

\item{skipvis, }{Set to True to skip visualizations}

\item{nCluster, }{number of interested cluster used for analysis}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
folders with prefix output. More info in output at https://hgmd.readthedocs.io/en/latest/Output.html
}
\description{
This function executes a ubuntu docker for cometsc (https://github.com/MSingerLab/COMETSC)
}
\examples{
\dontrun{
    #running cometsc
    cometsc(group="docker", file="/Users/raffaelecalogero/Desktop/AXLN1/data/topx_veanno.csv", 
           scratch.folder="/Users/raffaelecalogero/Desktop",
           threads=1, counts="True", skipvis="False", nCluster=8, separator=",") 
}

}
\author{
Raffaele Calogero,raffaele.calogero [at] unito [dot] it, University of Torino
}
