% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2log.R
\name{counts2log}
\alias{counts2log}
\title{A function to convert raw count in log counts}
\usage{
counts2log(counts.matrix, data.folder = getwd(), log.base = c(2, 10),
  type = c("txt", "csv"))
}
\arguments{
\item{counts.matrix, }{a character string indicating the name of the input raw counts file.}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{log.base, }{the base of the log to be used for the transformation}

\item{type, }{the type of input file, txt, tab delimited. csv, comma separated}
}
\value{
log transformed table
}
\description{
This function convert raw counts in log format
}
\examples{
\dontrun{
    system("wget http://130.192.119.59/public/TO BE INSERTED")
    #running skeleton
    counts2log(group="docker", counts.matrix, data.folder=getwd(), log.base=10, type="txt")
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino
}
