% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvToSparse.R
\name{csvToSparse}
\alias{csvToSparse}
\title{csv to Sparse}
\usage{
csvToSparse(group = c("sudo", "docker"), scratch.folder, file, separator)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
three files: matrix.mtx, barcodes.tsv and genes.tsv
}
\description{
This function executes a ubuntu docker that create a sparse matrix, without 0s, from a dense one, including 0s
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")
csvToSparse(group="docker", scratch="/data/scratch", 
            file=paste(getwd(), "annotated_setPace_10000_noC5.txt", sep="/"), 
            separator="\\t")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
