% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepClustering.R
\name{deepClustering}
\alias{deepClustering}
\title{deepClustering}
\usage{
deepClustering(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  bias,
  finetune_iters = 1e+05,
  nEpochs = 50000,
  seed = 1111,
  projectName,
  bN = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{number of cluster in which the dataset is divided}

\item{bias, }{bias method to use : "mirna" , "TF", "CUSTOM", kinasi,immunoSignature, cytoBands, ALL}

\item{finetune_iters, }{number of deep iteration}

\item{nEpochs, }{number of Epochs for neural network training}

\item{seed, }{important value to reproduce the same results with same input}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{bN, }{name of the custom bias file. This file need header, in the first column has to be the source and in the second column the gene symbol. All path needs to be provided.}
}
\description{
The present function compress data using autoencoder partially connected
}
\examples{
\dontrun{
deepClustering(group=c("docker"), scratch.folder="/scratch/", file="/home/lucastormreig/clustering_autoencoder/DCA/Rfunction/setA2.csv",separator=",", nCluster=5, bias="TF", finetune_iters=3, nEpochs=2,seed=1111,projectName="TF",bN="NULL")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
