% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dePblkae.R
\name{dePblkae}
\alias{dePblkae}
\title{dePblkae}
\usage{
dePblkae(group = c("sudo", "docker"), data.folder, scratch.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{folder where the total.csv file is located}

\item{scratch.folder, }{where calculation is performed}
}
\value{
a file called DE.txt, which contains for each cluster the ID of the genes detected as DE,  a file called log2_reformatedCPM.psblkAE.csv, which contains the log2 CPM of the psblkAE file and a file log2_reformatedCPM.psblkAE_mean-centered.csv, where log2 CPM are mena centered.
}
\description{
This function execute dePblkae analysis which detects differentiale expressed genes among clusters using a pseudobulk generated with autoencoders: autoencoder4pseudoBulk function. It is advisible to  do the analysis on a dataset made of at least 20 permutation
}
\examples{
\dontrun{
 library(rCASC)
 dePblkae(group="docker", 
        data.folder="/somewhere/in/your/PC", 
        scratch.folder="/scratch"
 )
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
