% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{dimensions}
\alias{dimensions}
\title{Cells count table size}
\usage{
dimensions(group = c("sudo", "docker"), scratch.folder, file, separator)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
number of row and columns of a cells counts table
}
\description{
This function executes a ubuntu docker that counts row and colums of a counts table.
}
\examples{
\dontrun{
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
