% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterZeros.R
\name{filterZeros}
\alias{filterZeros}
\title{A function plotting the distribution of zeros in cells eliminating all genes without a certain fraction of 0 in the samples}
\usage{
filterZeros(file, threshold = 0, sep)
}
\arguments{
\item{file, }{a character string indicating the path of the cells un-normalized expression counts table}

\item{threshold, }{a number from 0 to 1 indicating the fraction of max accepted zeros in each gene. 0 is set as default and it eliminates only genes having 0 in all cells.}

\item{sep, }{separator used in count file, e.g. '\\t', ','}
}
\value{
a PDF providing zeros distributions before removal of all genes without a certain fraction of 0 in the samples. A file with the prefix *filtered* in which the filtered data are saved.
}
\description{
This function plots the zeros distributions in cells and removes genes without  a certain fraction of 0 in the samples
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    filterZeros(file=paste(getwd(),"singlecells_counts.txt",sep="/"), threshold=0.1, sep="\\t")
}
}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino
}
