% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlrGenesSelection.R
\name{genesSelection}
\alias{genesSelection}
\title{Genes selection from genesPrioritization output}
\usage{
genesSelection(group = c("sudo", "docker"), scratch.folder, file,
  nCluster, separator, seed = 111, sp = 0.8, clusterPermErr = 0.05,
  maxDeltaConfidence = 0.01, minLogMean = 0.05)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{the number of clusters, where to run prioritization}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{seed, }{important value to reproduce the same results with same input}

\item{sp, }{minimun number of percentage of cells that has to be in common in a cluster, between two permutations, default 0.8}

\item{clusterPermErr, }{probability error in depicting the number of clusters in each permutation, default = 0.05}

\item{maxDeltaConfidence, }{max value for Delta confidence for gene prioritization p-values.}

\item{minLogMean, }{min value for Log mean gene prioritization p-value. P-value indicates the importance of a gene in defining clusterization.}
}
\value{
....
}
\description{
This function executes a ubuntu docker that extracts the genes playin major role in clusterin from output of genesPrioritization
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/section4.1_examples.zip")
unzip("section4.1_examples.zip")
setwd("section4.1_examples")
genesSelection(group="docker",scratch.folder="/data/scratch/",file=paste(getwd(), "bmsnkn_5x100cells.txt", sep="/"), nCluster=5, separator="\\t",  seed=111, sp=0.8, clusterPermErr=0.05, maxDeltaConfidence=0.01, minLogMean=0.05)
}

}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
