% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneVisualization2.R
\name{geneVisualization2}
\alias{geneVisualization2}
\title{GeneVisualization2}
\usage{
geneVisualization2(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  clustering.output,
  geneList,
  separator,
  finalName,
  pvalueFile,
  threshold
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the count matrix}

\item{clustering.output, }{a character string indicating the path of the clustering.output matrix}

\item{geneList, }{a character string indicating the path of the geneList matrix (no header, no row names)}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{finalName, }{name used for plot.}

\item{pvalueFile, }{name for pvalue file.}

\item{threshold, }{threshold value.}
}
\value{
plot
}
\description{
This function executes a ubuntu docker that performs geneVisualization
}
\examples{
\dontrun{
dir.create("scratch")
geneVisualization2(group="docker", scratch.folder="/scratch", file="/example/log2cpm.csv",clustering.output="/example/log2cpm_fake_clustering.output.csv",geneList="/example/geneList2.csv",separator=",",finalName="yo",pvalueFile="/example/log2cpmfisher.csv",threshold=0.05)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
