% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneVisualizationSpatial.R
\name{geneVisualizationSpatial}
\alias{geneVisualizationSpatial}
\title{GeneVisualization}
\usage{
geneVisualizationSpatial(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  tissuePosition,
  geneList,
  separator,
  finalName
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the count matrix}

\item{tissuePosition, }{a character string indicating the path of the tissuePosition matrix}

\item{geneList, }{a character string indicating the path of the geneList matrix (no header, no row names)}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{finalName, }{name used for plot.}
}
\value{
plot
}
\description{
This function executes a ubuntu docker that performs geneVisualization
}
\examples{
\dontrun{
dir.create("scratch")
geneVisualization(group=c("sudo"), scratch.folder=paste(getwd(),"/scratch",sep=""), file=paste(getwd(),"/setA.csv",sep=""),tissuePosition=paste(getwd(),"/setA_tissuePosition.csv",sep=""),geneList=paste(getwd(),"/geneList.csv",sep=""),separator=",",finalName="Lista1")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
