% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbscluster.R
\name{gibbsR}
\alias{gibbsR}
\title{gibbsR}
\usage{
gibbsR(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  resFolderCustom = "NULL",
  jobName,
  nCluster,
  motifLength,
  maxDelLength,
  maxInsLength,
  numbOfSeed,
  penalityFactorIntCluster = 0.8,
  backGroundAminoFreq = 2,
  seqWeightType = 1
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{path to peptide file,}

\item{resFolderCustom, }{optional parameter. Default will store the results in fastqPath otherwise will store the results in resFolderCustom path.}

\item{jobName, }{identify of the sample}

\item{nCluster, }{number of clusters to be generated}

\item{motifLength, }{length of the motif to be searched}

\item{maxDelLength, }{max length of deletions in the peptide search, expressed in number of aa}

\item{maxInsLength, }{max length of insertion in the peptide search, expressed in number of aa}

\item{numbOfSeed, }{number of initial configurations to be used to detect the highest KLD sum}

\item{penalityFactorIntCluster, }{default 0.8,}

\item{backGroundAminoFreq, }{two value 1 defined by uniprot and 2 defined by the dataset under analysis}

\item{seqWeightType, }{sequence weighting type: default 1, faster but less precise, 2 slower but more precise, 3 none}
}
\value{
an indexed genome compliant with 10XGenomics cellranger
}
\description{
This function creates performs motif search in peptides. It was developed to use gibbscluster on TCR CDR3 peptide sequences
}
\examples{
\dontrun{
library(rCASC)
dir.create("scratch")
scratch.folder=paste(getwd(),"scratch",sep="/")
fastqPath=paste(getwd(),"fastq",sep="/")
resFolder=paste(getwd(),"resFolder",sep="/")
dir.create(resFolder)

gibbsR(group="docker",scratch.folder=scratch.folder,file=file,resFolderCustom=results,jobName="test",nCluster=8,motifLength=4,maxDelLength=4,maxInsLength=4,numbOfSeed=4,penalityFactorIntCluster=0.8,backGroundAminoFreq=2,seqWeightType=1) 

newFolder=paste(results,"res",list.files(paste(results,"res",sep="/")),"cores",sep="/")
}


}
\author{
Luca Alessandrì
}
