% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griphBootstrap.R
\name{griphBootstrap}
\alias{griphBootstrap}
\title{Permutations and Clustering}
\usage{
griphBootstrap(group = c("sudo", "docker"), scratch.folder, file, nPerm,
  permAtTime, percent, separator, logTen = 0, perplexity = 10,
  seed = 111, format = NULL)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to be executed}

\item{permAtTime, }{number of permutations computed in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{perplexity, }{perplexity value for tsne projection}

\item{seed, }{important value to reproduce the same results with same input}

\item{format, }{output file format csv or txt}
}
\value{
VioPlot of silhouette cells value for each number of cluster used,clusterP file with clustering results for each permutation, killedCell file with removed cells in each permutation, clustering.output a sommarize file with general information for each cells.
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering.
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/section4.1_examples.zip")
unzip("section4.1_examples.zip")
setwd("section4.1_examples")
griphBootstrap(group="docker",scratch.folder="/data/scratch/",file=paste(getwd(), "bmsnkn_5x100cells.txt", sep="/"), nPerm=160, permAtTime=8, percent=10, separator="\\t",logTen=0, pca_number=6, seed=111, sparse=FALSE,format="NULL")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
