% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5tocsv.R
\name{h5tocsv}
\alias{h5tocsv}
\title{sparse to dense}
\usage{
h5tocsv(group = c("sudo", "docker"), file, type = c("h5", "10xgenomics"))
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{file, }{path of the sparse matrix file. For h5 file the full path MUST be included. For mtx matrix the folder MUST contain tsv and mtx files and the FULL path to mtx matrix MUST be provided}

\item{type, }{h5 refers to h5 files and 10xgenomics to the folder containing barcodes.tsv, genes.tsv and matrix.mtx}
}
\value{
a dense matrix, with 0s, in csv and txt format
}
\description{
This function takes h5 or mtx file from cellranger output and convert it in a csv table.
}
\examples{
\dontrun{
home <- getwd()
library(rCASC)
#download from https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE106268 the GSE106268_RAW.tar file
system("tar xvf GSE106268_RAW.tar")
h5tocsv(group="docker", file=paste(getwd(),"GSM2833284_Naive_WT_Rep1.h5",sep="/"), type="h5"))
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")
csvToSparse(group="docker", scratch="/data/scratch", file=paste(getwd(), 
            "annotated_setPace_10000_noC5.txt", sep="/"), separator="\t")
h5tocsv(group="docker", file=paste(getwd(),"matrix.mtx",sep="/"), type="10xgenomics")
}


}
\author{
Raffaele Calogero, raffaele [dot] calogero [at] unito [dot] com, University of Torino
}
