% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmony}
\alias{harmony}
\title{Harmony}
\usage{
harmony(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  meta_data,
  usePCA
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included. The file must contain library normalized cells, log transformed counts, and scaled genes}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{meta_data, }{file with meta_data. All path included.Check example file.}

\item{usePCA, }{TRUE or FALSE if Harmony has to use pca or not.}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function performs integration of single cell genomics datasets. Korsunsky et al. Nature Methods 16, 1289, 2019
}
\examples{
\dontrun{
harmony(group="sudo", scratch.folder="/home/lucastormreig/newDockers/harmony/scratch/", file="/home/lucastormreig/newDockers/harmony/scratch/example.csv",separator=",",meta_data="/home/lucastormreig/newDockers/harmony/scratch/example_metaData.csv",usePCA=FALSE)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
