% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfc.R
\name{hfc}
\alias{hfc}
\title{hfc, heatmap for clustering}
\usage{
hfc(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nCluster,
  separator,
  lfn,
  geneNameControl = 1,
  status,
  b1 = 0,
  b2 = 0
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{nCluster interested for the analysis}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{lfn, }{name of the list of genes}

\item{geneNameControl, }{0 if the matrix has gene name without ENSEMBL geneID. 1 if the gene names is formatted as: ENSMUSG00000000001:Gnai3. If the gene names is made only by ensamble name, scannoByGtf has to be run first.}

\item{status, }{0 if is raw count, 1 otherwise}

\item{b1, }{the lower range of signal in the heatmap,for negative value write "/-5". To ask the function to do automatic value assignment set 0}

\item{b2, }{the upper range of signal in the heatmap,for negative value write "/-2". To ask the function to do automatic value assignment set 0}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function generate heatmap and other plot based on clustering and on a specific gene list
}
\examples{
\dontrun{
hfc("docker","/home/lucastormreig/scratch/","/home/lucastormreig/CASC7.2/6_1hfc/Data/random_10000_filtered_annotated_lorenz_naive_penta2_0",6,",","naive")#
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
