% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indropIndex.R
\name{indropIndex}
\alias{indropIndex}
\title{A function to create a genome index for indrop V2 single cell data}
\usage{
indropIndex(group = c("sudo", "docker"), index.folder, ensembl.urlgenome,
  ensembl.urlgtf)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{index.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{ensembl.urlgenome, }{a character string indicating the URL from ENSEMBL ftp for the unmasked genome sequence of interest}

\item{ensembl.urlgtf, }{a character string indicating the URL from ENSEMBL ftp for the GTF for genome of interest}
}
\description{
This function executes a docker that produces as output the single cell index for V2 indrop single cell sequencing
}
\examples{
\dontrun{
library(docker4seq)
#running indropCounts index build
indropIndex(group="docker", index.folder="/data/genomes/mm10indrop",
    ensembl.urlgenome="ftp://ftp.ensembl.org/pub/release-87/fasta/mus_musculus/dna/Mus_musculus.GRCm38.dna.toplevel.fa.gz",
    ensembl.urlgtf="ftp://ftp.ensembl.org/pub/release-87/gtf/mus_musculus/Mus_musculus.GRCm38.87.gtf.gz")
}

}
\author{
Raffaele Calogero and Riccardo Panero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
