% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrationCircos.R
\name{integrationCircos}
\alias{integrationCircos}
\title{integrationCircos}
\usage{
integrationCircos(
  group = c("sudo", "docker"),
  scratch.folder,
  gsea.file,
  isc.file,
  XYpb.file,
  pblkae.file,
  Xcls.groups = NULL,
  Ycls.groups = NULL,
  outputFolder
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{gsea.file, }{a character string indicating the path of the final_score.csv  generated with gseaXLmHG, file, with file name and extension included.}

\item{isc.file, }{a character string indicating the path of the final_score.csv  generated with seuratIntegrationPermutation, file, with file name and extension included. included.}

\item{XYpb.file, }{a character string indicating the path of the final_score.csv  generated with integrationPsblk, file, with file name and extension included.}

\item{pblkae.file, }{a character string indicating the path of the final_score.csv  generated with integrationPblkae, file, with file name and extension included.}

\item{Xcls.groups, }{a vector of strings describing the groups of more similar clusters.  The optimal order of the clusters can be deduced by the output of integrationPblkae function with the option type="intra". Format: c("1cl6", "1cl2-1cl3", "1cl1-1cl4-1cl5").}

\item{Ycls.groups, }{a vector of strings describing the order of the clusters.  The optimal order of the clusters can be deduced by the output of integrationPblkae function with the option type="intra".  Format: c("2cl1-2cl3", "2cl2-2cl4", "2cl5-2cl6")}

\item{outputFolder, }{where results are placed}
}
\value{
A picture called integrated_score.png and a file called integrated_score.csv and all the final_scores.csv used to produce the integrated results. The colour ramp for Xgroups is yellow-magenta as instead for the Ygroups is green-blue. Clusters sharing the same colors are those characterized by belonging to the same subgroup detected using integrationPblkae with type="intra". Picture and data used for the integration are lccated in integrated_score folder
}
\description{
This function executes integrationCircos, which plots the integrated results of the analysis performed with gseaXLmHG, seuratIntegrationPermutation, integrationPsblk, integrationPblkae and BC
}
\examples{
\dontrun{
library(rCASC)
integrationCircos(group="docker", 
                  scratch.folder="/scratch", 
                  gsea.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/GSEA/final_score.csv",
                  isc.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/ISC/final_score.csv",
                  XYpb.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/XYpb/XYpb_final_score.csv",
                  pblkae.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/pblkAE/final_score.csv",
                  Xcls.order=NULL,
                  Ycls.order=NULL, 
                  outputFolder="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2"
)

integrationCircos(group="docker", 
                  scratch.folder="/scratch", 
                  gsea.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/GSEA/final_score.csv",
                  isc.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/ISC/final_score.csv",
                  XYpb.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/XYpb/XYpb_final_score.csv",
                  pblkae.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/pblkAE/final_score.csv",
                  Xcls.groups=c("1cl6", "1cl2-1cl3", "1cl1-1cl4-1cl5"),
                  Ycls.groups=c("2cl1-2cl3", "2cl2-2cl4", "2cl5-2cl6"), 
                  outputFolder="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2"
)

}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
