% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrationPblk.R
\name{integrationPsblk}
\alias{integrationPsblk}
\title{integrationPsblk}
\usage{
integrationPsblk(
  group = c("sudo", "docker"),
  scratch.folder,
  fileX,
  fileY,
  separatorX,
  separatorY,
  max.genes = 500,
  outputFolder
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{fileX, }{a character string indicating the path of the pseudobulkRow file, with file name and extension included.}

\item{fileY, }{a character string indicating the path of the pseudobulkRow file, with file name and extension included.}

\item{separatorX, }{separator used in count file, e.g. '\\t', ','}

\item{separatorY, }{separator used in count file, e.g. '\\t', ','}

\item{max.genes, }{MAX number of random genes to be used for each cluster, default 500}

\item{outputFolder, }{where results are placed}
}
\value{
A folder called XYpb with all the results generated. the final frequence table is saved in final_score.csv. The function produces an integrated output combining 7 thresholds from max.genes (top.ranked, top.ranked/2, top.ranked/4, top.ranked/8, top.ranked/16/ top.ranked/32, top.ranked/64). Pearson correlation is calculated on 1000 random selections of genes for each threshold. it is returned the frequence of having pearson >= 0.5.
}
\description{
This function execute integrationPsblk analysis which search for correspondence between clusters of two different experiments using clusters-pseudobulks, zscored on rows, and a subset of randomly selexte genes. Thus, the function clustersBulk has to be run in the two datasets before their comparison.
}
\examples{
\dontrun{
 library(rCASC)
 integrationPsblk(group="docker", 
        scratch.folder="/scratch", 
        fileX="/data/reanalysis_on_AIsc/comparing_CRC0327/NT_CTX/CRC0327_NT_2_clx/VandE/VandE_bulkRow.csv",
        fileY="/data/reanalysis_on_AIsc/comparing_CRC0327/NT_CTX/CRC0327_cetux_2_clx/VandE/VandE_bulkRow.csv", 
        separatorX=",",
        separatorY=",",
        max.genes=320,
        outputFolder="/data/reanalysis_on_AIsc/comparing_CRC0327/NT_CTX"
 )
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
