% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenzFilter.R
\name{lorenzFilter}
\alias{lorenzFilter}
\title{A function to handle sigle cell Lorenz Quality filter for Single-cells}
\usage{
lorenzFilter(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  p_value,
  separator
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file. IMPORTANT: full path to the file MUST be included}

\item{p_value, }{lorenz statistics threshold, suggest value 0.05, i.e. 5\% probability that the cell of low quality is selected}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
The output is a filtered counts table. Output will be in the same format and with the same separator of input.
}
\description{
This function executes a docker that embeds Lorenz filter from Diaz at al. Bioinformatics 2016
}
\examples{
\dontrun{
        system("wget http://130.192.119.59/public/testSCumi_mm10.csv.zip")
     library(casc)
     system("unzip testSCumi_mm10.csv.zip")
     #filtering low quality cells
     lorenzFilter(group="docker",scratch.folder="/data/scratch/", 
                  file=paste(getwd(),"testSCumi_mm10.csv", sep="/"),
                  p_value=0.05, separator=',')
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
