% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMatrix.R
\name{mergeMatrix}
\alias{mergeMatrix}
\title{Merge matrix}
\usage{
mergeMatrix(group = c("sudo", "docker"), scratch.folder, file1, file2,
  separator1, separator2, name1, name2)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file1, }{a character string indicating the path of the file, with file name and extension included}

\item{file2, }{a character string indicating the path of the file, with file name and extension included}

\item{separator1, }{separator used in count file, e.g. '\\t', ','}

\item{separator2, }{separator used in count file, e.g. '\\t', ','}

\item{name1, }{the name that will be in the header of the matrix1 cells ,leave to NULL if you dont want to put anything else}

\item{name2, }{the name that will be in the header of the matrix2 cells ,leave to NULL if you dont want to put anything else}
}
\value{
plot with cluster stability
}
\description{
This function executes a ubuntu docker that merge two matrix
}
\examples{
\dontrun{
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
