% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitoriboumi.R
\name{mitoRiboUmi}
\alias{mitoRiboUmi}
\title{MitoRiboUmi}
\usage{
mitoRiboUmi(group = c("sudo", "docker"), scratch.folder, file, separator,
  gtf.name, bio.type, umiXgene)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{gtf.name, }{name for the gtf file tu be used.}

\item{bio.type, }{ENSEMBL biotype of interest, default for single-cell protein_coding}

\item{umiXgene, }{a integer defining how many UMI are required to call a gene as present. default: 3}
}
\value{
a pdf file called Ribo_mito.pdf
}
\description{
This function executes a ubuntu docker showing ribosomal mitochondrial protein genes fraction in each cell
}
\examples{
\dontrun{
   system("wget ftp://ftp.ensembl.org/pub/release-94/gtf/mus_musculus/Mus_musculus.GRCm38.94.gtf.gz")
   system("gzip -d Mus_musculus.GRCm38.94.gtf.gz")
   system("wget http://130.192.119.59/public/testSCumi_mm10.csv.zip")
   unzip("testSCumi_mm10.csv.zip")
   mitoRiboUmi(group="docker", file=paste(getwd(), "testSCumi_mm10.csv", sep="/"), 
                scratch.folder="/data/scratch", separator=",", umiXgene=3, 
                gtf.name="Mus_musculus.GRCm38.94.gtf", bio.type="protein_coding")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
