% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodels.R
\name{mixmodels}
\alias{mixmodels}
\title{mixModels}
\usage{
mixmodels(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  geneList,
  separator,
  k,
  maxit = 10000
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the count matrix}

\item{geneList, }{a character string indicating the path of the geneList matrix (no header, no row names)}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{k, }{number of soft labels to divide the density}

\item{maxit, }{number of max iteration}
}
\value{
plot
}
\description{
This function executes a ubuntu docker that performs mixModels algorithm
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
