% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nClusterEvaluationSIMLR.R
\name{nClusterEvaluationSIMLR}
\alias{nClusterEvaluationSIMLR}
\title{Evaluation of the numbur of optimal clusters}
\usage{
nClusterEvaluationSIMLR(group = c("sudo", "docker"), scratch.folder,
  file, range1, range2, separator, logTen, nPerm, percent, rangeFile)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{range1, }{First number of cluster that has to be analyzed}

\item{range2, }{Last number of cluster that has to be analyzed}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{0 for raw count, 1 otherwise}

\item{nPerm, }{how many times number of cluster will be evaluated}

\item{percent, }{number of cells that has to be removed for bootstrap}

\item{rangeFile, }{name of the file that has for each row each number of cluster that has to be analyzed}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function analyze the data that came up from permutationClustering script.
}
\examples{
\dontrun{
permAnalysis("docker","path/to/scratch","path/to/data/TOTAL",3,4,",",0.8)#
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
