% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permAnalysis.R
\name{permAnalysis}
\alias{permAnalysis}
\title{Permutation Analysis}
\usage{
permAnalysis(group = c("sudo", "docker"), scratch.folder, data.folder,
  matrixName, range1, range2, format, separator, sp, clusterPermErr = 0.05)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{range1, }{First number of cluster that has to be analyzed}

\item{range2, }{Last number of cluster that has to be analyzed}

\item{format, }{matrix count format, "csv", "txt"}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{sp, }{minimun number of percentage of cells that has to be in common between two permutation to be the same cluster.}

\item{clusterPermErr, }{error that can be done by each permutation in cluster number depicting.Default = 0.05}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function analyze the data that came up from permutationClustering script.
}
\examples{
\dontrun{
 system("wget http://130.192.119.59/public/permutationanalysis.zip")
 unzip("permutationanalysis.zip")
 setwd("./permutationanalysis)
 library("CASC")
 permAnalysis("docker", scratch.folder="/data/scratch, data.folder=getwd(),
     matrixName="lorenz_Buettner", range1=4, range2=5, format="csv",
     separator=",", sp=0.8, clusterPermErr=0.05)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
