% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permAnalysisSeurat.R
\name{permAnalysisSeurat}
\alias{permAnalysisSeurat}
\title{Permutation Analysis Seurat}
\usage{
permAnalysisSeurat(group = c("sudo", "docker"), scratch.folder, file,
  nCluster, separator, sp)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{number of cluster that has to be analyzed}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{sp, }{minimun number of percentage of cells that has to be in common between two permutation to be the same cluster.}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function analyze the data that came up from permutationClustering script.
}
\examples{
\dontrun{
permAnalysis("docker","path/to/scratch","path/to/data/TOTAL",3,4,",",0.8)#
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
