% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationClusteringSHARP.R
\name{permutationClusteringSHARP}
\alias{permutationClusteringSHARP}
\title{Permutations and Clustering}
\usage{
permutationClusteringSHARP(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nPerm,
  permAtTime,
  percent,
  separator,
  logTen = 0,
  seed = 1111,
  perplexity = 10
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{seed, }{important value to reproduce the same results with same input}

\item{perplexity, }{number of close neighbors for each point. This parameter is specific for tSne. Default value is 10.  the performance of t-SNE is fairly robust under different settings of the perplexity. The most appropriate value depends on the density of your data.  A larger/denser dataset requires a larger perplexity. Typical values for the perplexity range between 5 and 50}
}
\value{
VioPlot of silhouette cells value for each number of cluster used,clusterP file with clustering results for each permutation, killedCell file with removed cells in each permutation, clustering.output a sommarize file with general information for each cells.
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering.
}
\examples{
\dontrun{
 permutationClusteringSHARP(group="docker",scratch.folder="/home/lucastormreig/CASC2.0/permutationClustering/scratch/",file="/home/lucastormreig/CASC2.0/permutationClustering/Data/TOTAL.csv",nPerm=4,permAtTime=2,percent=10,3,4,separator=",",logTen=0, perplexity, seed=1111)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
