% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recatPrediction.R
\name{recatPrediction}
\alias{recatPrediction}
\title{Cell Cycle}
\usage{
recatPrediction(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  geneNameControl = 0,
  window = 1,
  seed = 111
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the folder where input data are located and where output will be written and matrix name "/bin/users/matrix.csv"}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{geneNameControl, }{0 if the matrix has gene name without ENSEMBL code. 1 if the gene names is formatted like this : ENSMUSG00000000001:Gnai3. If the gene names is only ensamble name you have to run SCannoByGtf before start using this script.}

\item{window, }{number  cells  collapsed in a single point of the lot. This make less noisy the output generated by the predictor, default 10.}

\item{seed, }{important parameter for reproduce the same result with the same input}
}
\value{
a pdf, called cellCycleRange.pdf, with predicted cell cycle behaviour
}
\description{
This function executes a ubuntu docker which associates a cell cycle state to each cell
}
\examples{
\dontrun{
#preparing the data for the analysis 
system("wget http://130.192.119.59/public/buettner_G1G2MS_counts.txt.zip")
unzip("buettner_G1G2MS_counts.txt.zip")

#annotating the data set to obtain the gene names in the format ensemblID:symbol 
scannobyGtf(group="docker", file=paste(getwd(),"buettner_G1G2MS_counts.txt",sep="/"),
gtf.name="Mus_musculus.GRCm38.94.gtf", biotype="protein_coding", 
mt=TRUE, ribo.proteins=TRUE,umiXgene=3, riboStart.percentage=0, 
riboEnd.percentage=100, mitoStart.percentage=0, mitoEnd.percentage=100, thresholdGenes=100)
#running cell cycle prediction
recatPrediction(group="docker",scratch.folder="/data/scratch",
                file=paste(getwd(), "annotated_buettner_G1G2MS_counts.txt", sep="/"), 
                separator="\t", geneNameControl=1, window=10, seed=111)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
