% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saver.R
\name{saver}
\alias{saver}
\title{Saver}
\usage{
saver(group = c("sudo", "docker"), scratch.folder, file, separator, ncores)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included. The file must contain library normalized cells, log transformed counts, and scaled genes}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{ncores, }{number of cores to use}
}
\value{
predicted matrix
}
\description{
This function performs Data normalization and prediction with SAVER
}
\examples{
\dontrun{
saver(group=c("sudo"), scratch.folder="/home/user/saver/inDocker/scratch/", file="/home/user/saver/inDocker/data/setA.csv",separator=",",ncores=16)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
