% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCannoByGtf.R
\name{SCannoByGtf}
\alias{SCannoByGtf}
\title{Annotating single cell counts table using ENSEMBL gtf and refGenome CRAN package}
\usage{
SCannoByGtf(group = c("docker", "sudo"), data.folder = getwd(),
  counts.table, gtf.name, biotype, mt = c(TRUE, FALSE), file.type = c("txt",
  "csv"))
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{data.folder, }{a character string indicating where counts file is located}

\item{counts.table, }{a character string indicating the counts table file}

\item{gtf.name, }{a character string indicating the ENSEMBL gtf file}

\item{biotype, }{a character string the biotypes of interest}

\item{mt, }{a boolean to define if mitocondrial genes have to be removed}

\item{file.type, }{type of file: txt tab separated columns csv comma separated columns}
}
\value{
one file: annotated_counts table, where ensembl ids are linked to gene symbols
}
\description{
This function executes the docker container annotate.1, where refGenome is used to annotate a single cell counts table with ensembl gene ids on first column using ENSEMBL GTF annotation
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/genes.results.gz")
    system("gzip -d genes.results.gz")
    #running rsemannoByGtf
    SCannoByGtf(group="docker", data.folder=getwd(), counts.table="GSM2833284_Naive_WT_Rep1.csv",
                  gtf.name="Mus_musculus.GRCm38.92.gtf",
                  biotype, mt=TRUE, file.type="csv")
}

}
\author{
Raffaele Calogero
}
