% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAtomic.R
\name{scAtomic}
\alias{scAtomic}
\title{scAtomic}
\usage{
scAtomic(group = c("sudo", "docker"), scratch.folder, file, separator)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
results
}
\description{
This function executes a ubuntu docker that execute scAtomic on a csv file
}
\examples{
\dontrun{
scAtomic(group="docker",scratch.folder="/home/lucastormreig/sc-atomic/wtc1cls1/",file="/home/lucastormreig/sc-atomic/wtc1cls1/s1abr_wt_c1_Cluster1.csv",separator=",")
}
}
\author{
Raffaele Calogero
}
