% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schcl.R
\name{schcl}
\alias{schcl}
\title{Assigning cell types}
\usage{
schcl(group = c("sudo", "docker"), file, separator)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{file, }{a character string indicating the folder where input data are located. The input file is the output of scannobyGtf.}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
return a table summarizing the the cell type association for each cluster
}
\description{
This function execute scHCL to assign cell types from Han et al  Nature 2020, 581:303-309
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/annotated_lorenz_testSCumi_mm10.csv.zip")
unzip("annotated_lorenz_testSCumi_mm10.csv.zip")
schcl(group="docker", file=paste(getwd(),"annotated_lorenz_testSCumi_mm10.csv", sep="/"), separator=",")
}
}
\author{
Raffaele Calogero, raffaele [dot] calogero [at] unito [dot] it, University of Torino
}
