% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_cCycle.R
\name{seurat_ccycle}
\alias{seurat_ccycle}
\title{Seurat cell cycle}
\usage{
seurat_ccycle(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  seed
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the first matrix}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{seed, }{integer file necessary for reproducibility}
}
\value{
file containing the cluster association in the datasets merged by seurat
}
\description{
This function executes a ubuntu docker that performs seurat cell cycle
}
\examples{
\dontrun{
library(rCASC)
source("seurat_cCycle.R")
path=getwd()
scratch=paste(path,"/scratch",sep="")
dir.create(scratch)
file1=paste(path,"/setA.csv",sep="")
seurat_ccycle(group="docker", scratch.folder=scratch, file= file1, separator=",",seed=1111)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
