% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIntegration.R
\name{seuratIntegration}
\alias{seuratIntegration}
\title{Seurat Integration}
\usage{
seuratIntegration(
  group = c("sudo", "docker"),
  scratch.folder,
  file1,
  file2,
  separator1,
  separator2,
  seed
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file1, }{a character string indicating the path of the first matrix}

\item{file2, }{a character string indicating the path of the second matrix}

\item{separator1, }{separator used in count file, e.g. '\\t', ','}

\item{separator2, }{separator used in count file, e.g. '\\t', ','}

\item{seed, }{integer file necessary for reproducibility}
}
\value{
file containing the cluster association in the datasets merged by seurat
}
\description{
This function executes a ubuntu docker that performs seurat integration
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
seuratIntegration(group=c("sudo"), scratch.folder="/home/user/scratch", file1="/home/user/dockerFile/Seurat_join_DAPUSHARE/function/example/set1.csv",file2="/home/user/dockerFile/Seurat_join_DAPUSHARE/function/example/setA.csv", separator1=",",separator2=",",seed=1111)
}
