% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIntegrationMulti.R
\name{seuratIntegrationMulti}
\alias{seuratIntegrationMulti}
\title{Seurat Integration}
\usage{
seuratIntegrationMulti(
  group = c("sudo", "docker"),
  scratch.folder,
  folder,
  separator,
  resName,
  format
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{folder, }{a character string indicating the path of the file folder}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{resName, }{name for result matrix}

\item{format, }{format for all the matrix that has to be integrated. Format has to be all the same}
}
\value{
file containing the normalized matrix with the anchors
}
\description{
This function executes a ubuntu docker that performs seurat integration
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
seuratIntegrationMulti(group=c("sudo"), scratch.folder="/home/user/scratch", folder="/home/user/folderWithFiles",separator=",",resName="test",format="csv")
}
