% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratIntegrationPermutation.R
\name{seuratIntegrationPermutation}
\alias{seuratIntegrationPermutation}
\title{Seurat Integration Permutation}
\usage{
seuratIntegrationPermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  file1,
  file2,
  separator1,
  separator2,
  cl1,
  cl2,
  permutation,
  seed,
  outputFolder,
  K = 0.8
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file1, }{a character string indicating the path of the first matrix}

\item{file2, }{a character string indicating the path of the second matrix}

\item{separator1, }{separator used in count file, e.g. '\\t', ','}

\item{separator2, }{separator used in count file, e.g. '\\t', ','}

\item{cl1, }{path of clustering.output for file1}

\item{cl2, }{path of clustering.output for file2}

\item{permutation, }{number of permutation for statistic}

\item{seed, }{integer file necessary for reproducibility}

\item{outputFolder, }{path to the output folder}

\item{K, }{resolution for seurat analysis}
}
\value{
A folder called ISC, the input data, the intermediate results and a comma separated file final_score.csv summarising the frequency by which the seuratIntegration function return a an integration cluster including at least 50% of the input clusters form X and Y datasets.
}
\description{
This function executes a ubuntu docker that performs seurat integration to identify associated clusters in two independent experiment. The analysis is repeated mutiple times removing 10% of the initial cells, to investigate how stable is the partitioning done with seuratIntegration function. This function requires that cells were clustered with any of the clustering tools included in rCASC.
}
\examples{
\dontrun{
seuratIntegrationPermutation(group="docker", scratch.folder="/home/user/scratch", file1="/home/user/dockerFile/Seurat_join_DAPUSHARE/function/example/set1.csv",file2="/home/user/dockerFile/Seurat_join_DAPUSHARE/function/example/setA.csv", separator1=",",separator2=",",cl1=, cl29,permutation=100, seed=111) 
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
