% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratprior.R
\name{seuratPrior}
\alias{seuratPrior}
\title{seuratprior}
\usage{
seuratPrior(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  logTen = 0,
  seed = 111,
  PCADim,
  geneNumber,
  nCluster,
  sparse = FALSE,
  format = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{seed, }{important value to reproduce the same results with same input}

\item{PCADim, }{dimensions of PCA for seurat clustering}

\item{geneNumber, }{numbers of specific genes for each clusters}

\item{nCluster, }{number of cluster analysis.}

\item{sparse, }{boolean for sparse matrix}

\item{format, }{output file format}
}
\value{
....
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering.
}
\examples{
\dontrun{
 system("wget http://130.192.119.59/public/section4.1_examples.zip")
 unzip("section4.1_examples.zip")
 setwd("section4.1_examples")
 system("wget ftp://ftp.ensembl.org/pub/release-94/gtf/homo_sapiens/Homo_sapiens.GRCh38.94.gtf.gz")
 system("gzip -d Homo_sapiens.GRCh38.94.gtf.gz")
 system("mv Homo_sapiens.GRCh38.94.gtf genome.gtf")
 scannobyGtf(group="docker", file=paste(getwd(),"bmsnkn_5x100cells.txt",sep="/"),
 gtf.name="genome.gtf", biotype="protein_coding", mt=TRUE, ribo.proteins=TRUE,umiXgene=3)
 
 seuratPrior(group="docker", scratch.folder="/data/scratch/", file=paste(getwd(), "annotated_bmsnkn_5x100cells.txt", sep="/"), separator="\t", logTen=0, seed=111, PCADim=6, geneNumber = 100, nCluster=5)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
