% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlrBootstrap.R
\name{simlrBootstrap}
\alias{simlrBootstrap}
\title{Permutations and Clustering}
\usage{
simlrBootstrap(group = c("sudo", "docker"), scratch.folder, file, nPerm,
  permAtTime, percent, range1, range2, separator, logTen = 0, seed = 111,
  sp = 0.8, clusterPermErr = 0.05, maxDeltaConfidence = NULL,
  minLogMean = NULL)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to be executed}

\item{permAtTime, }{number of permutations computed in parallel}

\item{percent, }{percentage of cells retained in each permutation}

\item{range1, }{beginning of the range of clusters to be investigated}

\item{range2, }{end of the range of clusters to be investigated}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{seed, }{important value to reproduce the same results with same input}

\item{sp, }{minimun number of percentage of cells that has to be in common in a cluster, between two permutations, default 0.8}

\item{clusterPermErr, }{probability error in depicting the number of clusters in each permutation, default = 0.05}

\item{maxDeltaConfidence, }{max value for Delta confidence for genes feature selection}

\item{minLogMean, }{min value for Log mean for genes feature selection}
}
\value{
VioPlot of silhouette cells value for each number of cluster used,clusterP file with clustering results for each permutation, killedCell file with removed cells in each permutation, clustering.output a sommarize file with general information for each cells.
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering.
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/section4.1_examples.zip")
unzip("section4.1_examples.zip")
setwd("section4.1_examples")
simlrBootstrap(group="docker",scratch.folder="/data/scratch/",file=paste(getwd(), "bmsnkn_5x100cells.txt", sep="/"), nPerm=160, permAtTime=8, percent=10, range1=4, range2=6, separator="\\t",logTen=0, seed=111, sp=0.8, clusterPermErr=0.05)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
