% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slogos.R
\name{slogosR}
\alias{slogosR}
\title{slogosR}
\usage{
slogosR(
  group = c("sudo", "docker"),
  scratch.folder,
  gibbsCoresPath,
  resFolderCustom = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{gibbsCoresPath, }{Path to the cores folder that is generated as part of the results of gibbsR function.}

\item{resFolderCustom, }{optional parameter. Default will store the results in fastqPath otherwise will store the results in resFolderCustom path.}
}
\value{
an indexed genome compliant with 10XGenomics cellranger
}
\description{
This function creates slogs plot on the basis of the output of gibbsR function
}
\examples{
\dontrun{
library(rCASC)
dir.create("scratch")
scratch.folder=paste(getwd(),"scratch",sep="/")
gibbsCoresPath=paste(getwd(),"fastq",sep="/")
resFolder=paste(getwd(),"resFolder",sep="/")
dir.create(resFolder)
newFolder=paste(results,"res",list.files(paste(results,"res",sep="/")),"cores",sep="/")
slogosR(group="docker",scratch.folder,newFolder,resFolderCustom=newFolder)
}


}
\author{
Luca Alessandrì
}
