% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaGCNPermutation.R
\name{spaGCNPermutation}
\alias{spaGCNPermutation}
\title{spaGCN Permutation}
\usage{
spaGCNPermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  h5matrix.name,
  spotpositions.name,
  image.name,
  use_histology = TRUE,
  lResolution = 0.8,
  pcaDimensions = 10,
  p = 0.5,
  nPerm = 80,
  permAtTime = 8,
  percent = 10,
  seed = 1111
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{h5matrix.name, }{a character string indicating the full path + name with extension (.h5) of the expression matrix}

\item{spotpositions.name, }{a character string indicating the full path + name with extension (.csv) of the spot positions}

\item{image.name, }{a character string indicating the full path + name with extension (.tif) of the tissue image}

\item{use_histology, }{a boolean indicating if spaGCN should use the hisological information or not}

\item{lResolution, }{double for the resolution of Louvain algorithm}

\item{pcaDimensions, }{number of PCA dimension to keep}

\item{p, }{a real number between 0 and representing the percentage of total expression contributed by neighborhoods}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{seed, }{important value to reproduce the same results with same input}
}
\value{
To write
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering 
 using spaGCN
}
\author{
Giovanni Motterle, giovanni [dot] motterle [at] studenti [dot] univr [dot] it, University of Verona
}
