% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysis.R
\name{spatialAnalysis}
\alias{spatialAnalysis}
\title{spatial Analysis}
\usage{
spatialAnalysis(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nCluster,
  separator,
  tissuePosition
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{Last number of cluster that has to be analyzed}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{tissuePosition, }{file with tissue position name with extension}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function analyze the data that came up from permutationClustering script with spatial transcriptomics data.
}
\examples{
\dontrun{
spatialAnalysis(group="sudo", scratch.folder="/scratch", file="/home/lucastormreig/newDockers/data/matrix.csv",nCluster=13,separator=",",tissuePosition="/home/lucastormreig/newDockers/data/tissue_positions_list.csv")
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
