% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialAnalysisPlus.R
\name{spatialAnalysis2}
\alias{spatialAnalysis2}
\title{spatial Analysis Plus}
\usage{
spatialAnalysis2(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nCluster,
  separator,
  tissuePosition,
  Sp = 0.8,
  percentageIncrease = 10
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{Last number of cluster that has to be analyzed}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{tissuePosition, }{file with tissue position name with extension}

\item{Sp, }{Threshold to assign the plus score. 1 mean that all 6 cells, sorrounding the cell of interest, belong to the same cluster of the cell of interest. 0.83 out mean 5 out of 6 are of the same cluster of the reference cell, 0.66 mean 4 out of 6. We do not suggest to go below these three options.}

\item{percentageIncrease, }{percentage of the CSS score that has to be increased if the Threshold condition is satisfied.}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function analyzes the data that came up from permutationClustering script with spatial transcriptomics data. The function integrates CSS with a score, which is related to the beloging cluster identity of a cell with respect to the other 6 cell neighborhood that surround it.
}
\examples{
\dontrun{
spatialAnalysis2(group=c("sudo"), scratch.folder="/home/lucastormreig/temp/", file="/home/lucastormreig/test/spatialAnalysis/spatial/HLNrevised.csv",nCluster=5,separator=",",tissuePosition="/home/lucastormreig/test/spatialAnalysis/spatial/Results/tissue_positions_list.csv",Sp=0.8,percentageIncrease=10)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
