% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClusters.R
\name{splitClusters}
\alias{splitClusters}
\title{Split Matrix for each clusters}
\usage{
splitClusters(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{interesting number of clusters}
}
\value{
one Matrix for each cluster
}
\description{
This function executes a ubuntu docker that merge two matrix
}
\examples{
\dontrun{
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
