% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StardustMbkmeansPermutation.R
\name{StardustMbkmeansPermutation}
\alias{StardustMbkmeansPermutation}
\title{Stardust Mbkmeans Permutation}
\usage{
StardustMbkmeansPermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  tissuePosition,
  method = c("sw", "indsw"),
  spaceWeight = 1,
  center = 10,
  nPerm,
  permAtTime,
  percent,
  separator,
  logTen = 0,
  pcaDimensions = 10,
  seed = 1111,
  sparse = FALSE,
  format = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{tissuePosition, }{file with tissue position name with extension}

\item{method, }{method to be used to calculate the space weight. Two options: "sw" or "indsw", 
depending on whether the user specifies a weight or not}

\item{spaceWeight, }{double in [0,1]. Weight for the linear transormation of spot distance.
1 means spot distance weight as much as profile distance, 0 means spot distance doesn't 
contribute at all in the overall distance measure}

\item{center, }{number of centers for mini-batch k-means algorithm}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{pcaDimensions, }{0 for automatic selection of PC elbow}

\item{seed, }{important value to reproduce the same results with same input}

\item{sparse, }{boolean for sparse matrix}

\item{format, }{output file format csv or txt}
}
\value{
A specific number of permutations to evaluate clustering.
}
\description{
This function executes a ubuntu docker that produces a specific number of permutations to evaluate clustering 
 using an extended version of Seurat (updated to 2022) that takes into consideration physical position of spots.
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Eva Viesi, eva [dot] viesi [at] univr [dot] it, University of Verona

Simone Avesani, simone [dot] avesani [at] univr [dot] it, University of Verona
}
