% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartdustConfigurations.R
\name{StartdustConfigurations}
\alias{StartdustConfigurations}
\title{Stardust Configurations}
\usage{
StartdustConfigurations(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  tissuePosition,
  nPerm,
  permAtTime,
  percent,
  separator,
  logTen = 0,
  pcaDimensions = 30,
  seed = 1111,
  sparse = FALSE,
  format = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{tissuePosition, }{file with tissue position name with extension}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{pcaDimensions, }{0 for automatic selection of PC elbow.}

\item{seed, }{important value to reproduce the same results with same input}

\item{sparse, }{boolean for sparse matrix}

\item{format, }{output file format csv or txt}
}
\value{
5 results directories corrispondent to the 5 method configurations and a violin
 plots comparison of the CSSs of each configuration
}
\description{
This function executes the method StartdustPermutation 5
 times, each one with a different configuration for spaceWeight
 parameter (0,0.25,0.5,0.75,1). Requires ggplot2 installed.
}
\author{
Giovanni Motterle, giovanni [dot] motterle [at] studenti [dot] univr [dot] it, University of Verona
}
