% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLearnPermutation.R
\name{STLearnPermutation}
\alias{STLearnPermutation}
\title{StLearn Permutation}
\usage{
STLearnPermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  filtered_feature_bc_matrix,
  lResolution = 0.8,
  nPerm = 80,
  permAtTime = 8,
  percent = 10,
  pcaDimensions = 5,
  seed = 1111
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the full path + name with extension (.txt) of the expression matrix}

\item{filtered_feature_bc_matrix, }{a character string indicating the full path of the Space Ranger output directory (10X published data).
The directory must be named 'filtered_feature_bc_matrix'}

\item{lResolution, }{double for the resolution of Louvain algorithm}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{pcaDimensions, }{0 for automatic selection of PC elbow.}

\item{seed, }{important value to reproduce the same results with same input}
}
\value{
To write
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering 
 using seurat (updated to 2020) and a stlearn normalized expression matrix as input.
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Giovanni Motterle, giovanni [dot] motterle [at] studenti [dot] univr [dot] it, University of Verona
}
