% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toprnk.R
\name{toprnk}
\alias{toprnk}
\title{toprnk}
\usage{
toprnk(
  group = c("sudo", "docker"),
  scratch.folder,
  fileX,
  fileY,
  separatorX,
  separatorY,
  xCometFolder,
  yCometFolder,
  threshold = 0.8,
  top.ranked = 80,
  CSS.threshold = 0.7,
  outputFolder
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{fileX, }{a character string indicating the path of the pseudobulkRow file, with file name and extension included.}

\item{fileY, }{a character string indicating the path of the pseudobulkRow file, with file name and extension included.}

\item{separatorX, }{separator used in count file, e.g. '\\t', ','}

\item{separatorY, }{separator used in count file, e.g. '\\t', ','}

\item{xCometFolder, }{path of Comet results from X experiment}

\item{yCometFolder, }{path of Comet results from Y experiment}

\item{threshold, }{Pearson threshold}

\item{top.ranked, }{number of top comet genes to be used}

\item{CSS.threshold, }{min mean CSS that a cluster should have to be considered for integration}

\item{outputFolder, }{where results are placed}
}
\value{

}
\description{
This function execute toprnk analysis which search for corrispondence betwee clusters of two different experiments using pseudobulk zscored on rows and the cluster specific genes from comet analysis. Thsu the function clustersBulk and cometsc have to be run in the two dataset to be integrated.
}
\examples{
\dontrun{
 
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
