% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneBootstrap.R
\name{tsneBootstrap}
\alias{tsneBootstrap}
\title{Executing clustering with tSne}
\usage{
tsneBootstrap(group = c("sudo", "docker"), scratch.folder, file, nPerm,
  permAtTime, percent, range1, range2, separator, logTen = 0,
  seed = 111, sp = 0.8, clusterPermErr = 0.05, perplexity = 10)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to be executed}

\item{permAtTime, }{number of permutations computed in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{range1, }{beginning of the range of clusters to be investigated}

\item{range2, }{end of the range of clusters to be investigated}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{seed, }{important value to reproduce the same results with same input, default is 111}

\item{sp, }{minimun number of percentage of cells that has to be in common in a cluster, between two permutations, default 0.8}

\item{clusterPermErr, }{probability error in depicting the number of clusters in each permutation, default = 0.05}

\item{perplexity, }{number of close neighbors for each point. This parameter is specific for tSne. Default value is 10.  the performance of t-SNE is fairly robust under different settings of the perplexity. The most appropriate value depends on the density of your data.  A larger/denser dataset requires a larger perplexity. Typical values for the perplexity range between 5 and 50}
}
\value{
A folder Results containing a folder with the name of the experiment, which contains: VioPlot of silhouette cells value for each number of cluster used, a folder with the number of clusters used for SIMLR clustering, which contains: clusterP file with clustering results for each permutation, killedCell file with removed cells in each permutation, clustering.output a sommarize file with general information for each cells
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation using tSne as clustering tool.
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/section4.1_examples.zip")
unzip("section4.1_examples.zip")
setwd("section4.1_examples")
tsneBootstrap(group="docker",scratch.folder="/data/scratch/",file=paste(getwd(), "bmsnkn_5x100cells.txt", sep="/"), nPerm=160, permAtTime=8, percent=10, range1=4, range2=6, separator="\\t",logTen=0, seed=111, sp=0.8, clusterPermErr=0.05, perplexity=10)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
