% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap}
\alias{umap}
\title{umap}
\usage{
umap(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  seed,
  epochs
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the count matrix}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{seed, }{number for reproducibility}

\item{epochs, }{number of epochs for umap}
}
\value{
plot
}
\description{
This function executes a ubuntu docker that performs umap visualization
}
\examples{
\dontrun{
dir.create("scratch")
umap(group=c("sudo"), scratch.folder=paste(getwd(),"/scratch",sep=""), file=paste(getwd(),"/setA.csv",sep=""),separator=",",seed=111,epochs=1000)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
