% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updStardustPermutation.R
\name{updStardustPermutation}
\alias{updStardustPermutation}
\title{Updated Stardust Permutation}
\usage{
updStardustPermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  tissuePosition,
  res = 0.8,
  nPerm,
  permAtTime,
  percent,
  separator,
  logTen = 0,
  pcaDimensions = 5,
  seed = 1111,
  sparse = FALSE,
  format = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{tissuePosition, }{file with tissue position name with extension}

\item{res, }{double resolution for louvain algorithm}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{pcaDimensions, }{0 for automatic selection of PC elbow.}

\item{seed, }{important value to reproduce the same results with same input}

\item{sparse, }{boolean for sparse matrix}

\item{format, }{output file format csv or txt}
}
\value{
a specific number of permutation to evaluate clustering
}
\description{
This function executes a ubuntu docker that produces a specific number of permutations to evaluate clustering 
 using an extended version of seurat (updated to 2022) that take into consideration physical position of spots.
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Eva Viesi, eva [dot] viesi [at] univr [dot] it, University of Verona

Simone Avesani, simone [dot] avesani [at] univr [dot] it, University of Verona
}
