% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioHTMLbyGenes.R
\name{vioHTMLByGenes}
\alias{vioHTMLByGenes}
\title{vioHTMLByGenes}
\usage{
vioHTMLByGenes(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  cl,
  gene,
  logtwo
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{cl, }{Clustering.output file, which contains the output of every clustering algorithm embedded in rCASC or it can be a file having in the first column cells names and in the second column cluster to which the cell belongs. Full path is required.}

\item{gene, }{gene name that wants to be inspected. Gene format depends from the rows names format of the count file.}

\item{logtwo, }{1 if the matrix is already in log2, 0 otherwise}
}
\value{

}
\description{
The present function plot count matrix data into a vioplot for a gene in each cluster. I requires two file the count matrics and the _clustering.output file.
}
\examples{
\dontrun{
 vioHTMLByGenes(group=c("sudo"), scratch.folder="/home/lucastormreig/genesPlot/test/scratch", file="/home/lucastormreig/genesPlot/test/data/setA.csv",separator=",", cl="/home/lucastormreig/genesPlot/test/data/setA_clustering.output.csv",gene="ARF5",logtwo=1)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
