% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaLike.R
\name{anovaLike}
\alias{anovaLike}
\title{A function allowing the identification of differentially expressed genes.}
\usage{
anovaLike(group = c("sudo", "docker"), data.folder, counts.table,
  file.type = c("txt", "csv"), cluster.file, ref.cluster,
  logFC.threshold = 1, FDR.threshold, logCPM.threshold = 4, plot = c(TRUE,
  FALSE))
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{counts.table, }{a character string indicating the counts table file. IMPORTANT in the header of the file the covariate group MUST be associated to the column name using underscore, e.g. cell1_cov1}

\item{file.type, }{type of file: txt tab separated columns csv comma separated columns}

\item{cluster.file, }{a character string indicating the _clustering.output.txt file of interest, generated by bootstrapSimlar or bootStrapTsne}

\item{ref.cluster, }{a number indicating the cluster to be use da reference for comparison with the other clusters.}

\item{logFC.threshold, }{minimal logFC present in at least one of the comparisons with respect to reference covariate}

\item{FDR.threshold, }{minimal FDR present in at least one of the comparisons with respect to reference covariate}

\item{logCPM.threshold, }{minimal average abundance}

\item{plot, }{TRUE if differentially expressed genes are represented in a plot.}
}
\description{
This function executes in a docker edgeR for the idnetification of differentially expressed genes in single-cells RNAseq
}
\examples{
\dontrun{
    #running deDetection
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5_clustering.output.txt")
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")
anovaLike(group="docker", data.folder=getwd(), counts.table="annotated_setPace_10000_noC5_reordered.txt",
       file.type="txt", cluster.file="annotated_setPace_10000_noC5_clustering.output.txt", ref.cluster=3,
       logFC.threshold=1, FDR.threshold=0.05, logCPM.threshold=4, plot=TRUE)

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
