% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencoder}
\alias{autoencoder}
\title{Autoencoder}
\usage{
autoencoder(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  nCluster,
  bias,
  permutation,
  nEpochs,
  patiencePercentage = 5,
  cl,
  seed = 1111,
  projectName,
  bN = "NULL",
  lr = 0.01,
  beta_1 = 0.9,
  beta_2 = 0.999,
  epsilon = 1e-08,
  decay = 0,
  loss = "mean_squared_error",
  regularization = 10,
  variational = FALSE
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{nCluster, }{number of cluster in which the dataset is divided}

\item{bias, }{bias method to use : "mirna" , "TF", "CUSTOM", kinasi,immunoSignature,ALL}

\item{permutation, }{number of permutations to perform the pValue to evaluate clustering}

\item{nEpochs, }{number of Epochs for neural network training}

\item{patiencePercentage, }{number of Epochs percentage of not training before to stop.}

\item{cl, }{Clustering.output file. Can be the output of every clustering algorithm from rCASC or can be customized with first column cells names, second column cluster they belong. All path needs to be provided.}

\item{seed, }{important value to reproduce the same results with same input}

\item{projectName, }{might be different from the matrixname in order to perform different analysis on the same dataset}

\item{bN, }{name of the custom bias file. This file need header, in the first column has to be the source and in the second column the gene symbol. All path needs to be provided.}

\item{lr, }{learning rate, the speed of learning. Higher value may increase the speed of convergence but may also be not very precise}

\item{beta_1, }{look at keras optimizer parameters}

\item{beta_2, }{look at keras optimizer parameters}

\item{epsilon, }{look at keras optimizer parameters}

\item{decay, }{look at keras optimizer parameters}

\item{loss, }{loss of function to use, for other loss of function check the keras loss of functions.}

\item{regularization, }{this parameter balances between reconstruction loss and enforcing a normal distribution in the latent space.}

\item{variational, }{TRUE or FALSE if use the variational autoencoder or just the standard autoencoder.}
}
\description{
The present function compress data using autoencoder partially connected
}
\examples{
\dontrun{
 autoencoder(group="docker",scratch.folder="/home/user/Riccardo/Riccardo/1_inDocker/scratch",file="/home/user/Riccardo/Riccardo/1_inDocker/data/setA.csv",separator=",",nCluster=5,bias="TF",permutation=10,nEpochs=10,cl="/home/user/Riccardo/Riccardo/1_inDocker/data/setA_clustering.output.csv",projectName="testDocker")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
