% Generated by roxygen2 (4.0.1): do not edit by hand
\name{casc}
\alias{casc}
\title{Covariate Assisted Spectral Clustering}
\usage{
casc(adjMat, covMat, nBlocks, nIter = 10, method = "regLaplacian",
  rowNorm = TRUE)
}
\arguments{
\item{adjMat}{An adjacency matrix}

\item{covMat}{A covariate matrix}

\item{nBlocks}{The number of clusters}

\item{method}{The form of the adjacency matrix to be used.}

\item{rowNorm}{True if row normalization should be
done before running kmeans.}

\item{nIter}{Number of iterations to find  the optimal tuning
parameter.}
}
\value{
A list with node cluster assignments, the
the value of the tuning parameter used, the within
cluster sum of squares.
}
\description{
Covariate Assisted Spectral Clustering
}
\keyword{clustering}
\keyword{spectral}

