% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRemove.R
\name{ccRemove}
\alias{ccRemove}
\title{Cc remove}
\usage{
ccRemove(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  seed = 111,
  cutoff = 3,
  species,
  rawCount
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the folder where input data are located and where output will be written and matrix name "/bin/users/matrix.csv"}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{seed, }{is important to reproduce the same results with the same input}

\item{cutoff, }{a cut-off equal to 3 is almost equal to 0.05 p.value}

\item{species, }{human or mouse}

\item{rawCount, }{1 for raw 0 otherwise}
}
\value{
return a normalized matrix, whith prefix LS_cc_, in which cell cycle effect is removed
}
\description{
This function executes a ubuntu docker that remove the effect of cell cycle from a single cells dataset
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/annotated_lorenz_testSCumi_mm10.csv.zip")
unzip("annotated_lorenz_testSCumi_mm10.csv.zip")
ccRemove(group="docker" , scratch.folder="/data/scratch",
       file=paste(getwd(),"annotated_lorenz_testSCumi_mm10.csv", sep="/"), separator=",",
       seed=111, cutoff=3, species="mouse", rawCount=1)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
