% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCycle_1.R
\name{cellCycle1}
\alias{cellCycle1}
\title{Cell Cycle}
\usage{
cellCycle1(group = c("sudo", "docker"), scratch.folder, file, separator,
  geneNameControl = 0, window = 1, seed = 111)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the folder where input data are located and where output will be written and matrix name "/bin/users/matrix.csv"}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{geneNameControl, }{0 if the matrix has gene name without ENSEMBL code.1 if the gene names is formatted like this : ENSMUSG00000000001:Gnai3. If the gene names is only ensamble name you have to run SCannoByGtf before start using this script.}

\item{window, }{this parameters let you see the cell cycle algorithm plot in block.}

\item{seed, }{important parameter for reproduce the same result with the same input}
}
\value{
will change all the files generated from permAnalysis algorithm in a new folder matrixName_Cluster_merged/
}
\description{
This function executes a ubuntu docker that associates to each cell a cell cycle stage
}
\examples{
\dontrun{
#getwd(link)
#unzipFolder
scratch.folder=paste(getwd(),"/scratch",sep="")
file=paste(getwd(),"/data/annotated_Buettner.csv",sep="")
cellCycle1(group="docker",scratch.folder,file,separator=",",geneNameControl=1,window=10,seed=111)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
