% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCycle_2.R
\name{cellCycle2}
\alias{cellCycle2}
\title{Cell Cycle}
\usage{
cellCycle2(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  G1_a,
  G1_b,
  S_a,
  S_b,
  G2M_a,
  G2M_b,
  seed = 111
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the folder where input data are located and where output will be written and matrix name "/bin/users/matrix.csv"}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{G1_a, }{starting point in time series for G1 phase}

\item{G1_b, }{ending point in time series for G1 phase}

\item{S_a, }{starting point in time series for S phase}

\item{S_b, }{ending point in time series for S phase}

\item{G2M_a, }{starting point in time series for G2M phase}

\item{G2M_b, }{ending point in time series for G2M phase}

\item{seed, }{important parameter for reproduce the same result with the same input}
}
\value{
will change all the files generated from permAnalysis algorithm in a new folder matrixName_Cluster_merged/
}
\description{
This function executes a ubuntu docker that associates to each cell a cell cycle stage
}
\examples{
\dontrun{
#getwd(link)
#unzipFolder
scratch.folder=paste(getwd(),"/scratch",sep="")
file=paste(getwd(),"/data/annotated_Buettner.csv",sep="")
 cellCycle2(group="docker",scratch.folder,file,separator=",",G1_a=22,G1_b=23,S_a=22,S_b=23,G2M_a=85,G2M_b=86,seed=111)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
