% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_count.R
\name{cellranger_count}
\alias{cellranger_count}
\title{Cellranger count}
\usage{
cellranger_count(group = c("sudo", "docker"), id, transcriptome.folder,
  fastq.folder, sample, expect.cells = NULL, force.cells = NULL,
  nosecondary = FALSE, chemistry = NULL, r1.length = NULL,
  r2.length = NULL, lanes = NULL, localcores = NULL, localmem = NULL,
  indices = NULL, scratch.folder)
}
\description{
This function takes FASTQ files from cellranger mkfastq and performs alignment, filtering, barcode counting, and UMI counting. It uses the Chromium cellular barcodes to generate gene-barcode matrices, determine clusters, and perform gene expression analysis. The count pipeline can take input from multiple sequencing runs on the same library.
--id	     A unique run ID string: e.g. sample345
--fastq	 Either: Path of the fastq_path folder generated by cellranger mkfastq
          e.g. /home/jdoe/runs/HAWT7ADXX/outs/fastq_path. This contains a directory hierarchy that cellranger count will              #'           automatically traverse.
          - OR -
          Any folder containing fastq files, for example if the fastq files were generated by a service provider and delivered   #'           outside the context of the mkfastq output directory structure.
          Can take multiple comma-separated paths, which is helpful if the same library was sequenced on multiple flowcells.
          Doing this will treat all reads from the library, across flowcells, as one sample.
          If you have multiple libraries for the sample, you will need to run cellranger count on them individually, and then #'           combine them with cellranger aggr.
--sample	 Sample name as specified in the sample sheet supplied to cellranger mkfastq.
          Can take multiple comma-separated values, which is helpful if the same library was sequenced on multiple flowcells #'           and the sample name used (and therefore fastq file prefix) is not identical between them.
          Doing this will treat all reads from the library, across flowcells, as one sample.
          If you have multiple libraries for the sample, you will need to run cellranger count on them individually, and then #'           combine them with cellranger aggr.
          Allowable characters in sample names are letters, numbers, hyphens, and underscores.
--transcriptome       	Path to the Cell Ranger compatible transcriptome reference e.g.
                         For a human-only sample, use /opt/refdata-cellranger-GRCh38-1.2.0
                         For a human and mouse mixture sample, use /opt/refdata-cellranger-hg19-and-mm10-1.2.0
--expect-cells    (optional) Expected number of recovered cells. Default: 3,000 cells.
--force-cells	    (optional) Force pipeline to use this number of cells, bypassing the cell detection algorithm.
                 Use this if the number of cells estimated by Cell Ranger is not consistent with the barcode rank plot.
--nosecondary	    (optional) Add this flag to skip secondary analysis of the gene-barcode matrix (dimensionality reduction,
                 clustering and visualization). Set this if you plan to use cellranger reanalyze or your own custom analysis.
--chemistry	    (optional) Assay configuration. One of:
                 auto for autodetection (default), threeprime for Single Cell 3′, fiveprime for Single Cell 5′, SC3Pv1 for
                 Single Cell 3′ v1, SC3Pv2 for Single Cell 3′ v2, SC5P-PE for Single Cell 5′ paired-end (both R1 and R2 are
                 used for alignment),SC5P-R2 for Single Cell 5′ R2-only (where only R2 is used for alignment).
--r1-length	    (optional) Hard-trim the input R1 sequence to this length. Note that the length includes the Barcode and UMI
                 sequences so do not set this below 26 for Single Cell 3′ v2 or Single Cell 5′. This and --r2-length are
                 useful for determining the optimal read length for sequencing.
--r2-length	    (optional) Hard-trim the input R2 sequence to this length.
--lanes	        (optional) Lanes associated with this sample
--localcores	    Restricts cellranger to use specified number of cores to execute pipeline stages. By default, cellranger
                 will use all of the cores available on your system.
--localmem	    Restricts cellranger to use specified amount of memory (in GB) to execute pipeline stages. By default,
                 cellranger will use 90% of the memory available on your system. Please note that cellranger requires at least
                 16 GB of memory to run all pipeline stages.
--indices	        (Deprecated. Optional. Only used for output from cellranger demux) Sample indices associated with this
                 sample. Comma-separated list of:
                 index set plate well: SI-3A-A1
                 index sequences: TCGCCATA,GTATACAC
}
\examples{
\dontrun{
home <- getwd()
library(casc)
downloadContainers()
setwd("/data/genomes/cellranger_hg19mm10")
#getting the human and mouse cellranger index
system("wget http://cf.10xgenomics.com/supp/cell-exp/refdata-cellranger-hg19-and-mm10-2.1.0.tar.gz")
setwd(home)
# 100 cells 1:1 Mixture of Fresh Frozen Human (HEK293T) and Mouse (NIH3T3) Cells
system("wget http://cf.10xgenomics.com/samples/cell-exp/2.1.0/hgmm_100/hgmm_100_fastqs.tar")
cellranger_count(group="docker",  id="hgmm", transcriptome.folder="/data/genomes/cellranger_hg19mm10",  fastq.folder="/data/test_cell_ranger/fastqs",  sample="hgmm", expect.cells=100, nosecondary=TRUE, scratch.folder="/data/scratch")




}


}
\author{
????
}
