% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger_count.R
\name{cellranger_count}
\alias{cellranger_count}
\title{Cellranger count}
\usage{
cellranger_count(group = c("sudo", "docker"), transcriptome.folder,
  fastq.folder, sample = NULL, expect.cells = NULL, force.cells = NULL,
  nosecondary = FALSE, chemistry = NULL, r1.length = NULL,
  r2.length = NULL, lanes = NULL, localcores = NULL, localmem = NULL,
  scratch.folder)
}
\arguments{
\item{nosecondary, }{(optional) Add this flag to skip secondary analysis of the gene-barcode matrix (dimensionality reduction, clustering and visualization). Set this if you plan to use cellranger reanalyze or your own custom analysis.}

\item{chemistry, }{(optional) Assay configuration. One of: auto for autodetection (default), threeprime for Single Cell 3′, fiveprime for Single Cell 5′, SC3Pv1 for Single Cell 3′ v1, SC3Pv2 for Single Cell 3′ v2, SC5P-PE for Single Cell 5′ paired-end (both R1 and R2 are used for alignment),SC5P-R2 for Single Cell 5′ R2-only (where only R2 is used for alignment).}

\item{lanes, }{(optional) Lanes associated with this sample}

\item{localcores, }{restricts cellranger to use specified number of cores to execute pipeline stages. By default, cellranger will use all of the cores available on your system.}

\item{localmem, }{restricts cellranger to use specified amount of memory (in GB) to execute pipeline stages. By default, cellranger will use 90% of the memory available on your system. Please note that cellranger requires at least 16 GB of memory to run all pipeline stages.}

\item{fastq, }{path of the fastq_path folder}

\item{transcriptome, }{path to the Cell Ranger compatible transcriptome reference e.g. for a human and mouse mixture sample, use refdata-cellranger-hg19-and-mm10-1.2.0}

\item{expect-cells, }{(optional) Expected number of recovered cells. Default: 3,000 cells.}

\item{force-cells, }{(optional) Force pipeline to use this number of cells, bypassing the cell detection algorithm. Use this if the number of cells estimated by Cell Ranger is not consistent with the barcode rank plot.}

\item{r1-length, }{(optional) Hard-trim the input R1 sequence to this length. Note that the length includes the Barcode and UMI sequences so do not set this below 26 for Single Cell 3′ v2 or Single Cell 5′. This and --r2-length are useful for determining the optimal read length for sequencing.}

\item{r2-length, }{(optional) Hard-trim the input R2 sequence to this length.}
}
\value{
a folder called results_cellranger, more info on the structure of this folder at https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/overview
}
\description{
This function takes FASTQ files from cellranger mkfastq and performs alignment, filtering, barcode counting, and UMI counting. It uses the Chromium cellular barcodes to generate gene-barcode matrices, determine clusters, and perform gene expression analysis. The count pipeline can take input from multiple sequencing runs on the same library.
}
\examples{
\dontrun{
home <- getwd()
library(casc)
downloadContainers()
setwd("/data/genomes/cellranger_hg19mm10")
#getting the human and mouse cellranger index
system("wget http://cf.10xgenomics.com/supp/cell-exp/refdata-cellranger-hg19-and-mm10-2.1.0.tar.gz")
setwd(home)
# 100 cells 1:1 Mixture of Fresh Frozen Human (HEK293T) and Mouse (NIH3T3) Cells
system("wget http://cf.10xgenomics.com/samples/cell-exp/2.1.0/hgmm_100/hgmm_100_fastqs.tar")
cellranger_count(group="docker",  transcriptome.folder="/data/genomes/cellranger_hg19mm10",  fastq.folder="/data/test_cell_ranger/fastqs",  expect.cells=100, nosecondary=TRUE, scratch.folder="/data/scratch")
# This analysis took 56.4 mins on an Intel NUC6I7KYK with 32 Gb RAM, Intel i7-6770HQ 8 threads, 1Tb SSD.



}


}
\author{
Greta Romano, romano [dot] greta [at] gmail [dot] com, University of Torino
}
