% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCountDepth.R
\name{checkCountDepth}
\alias{checkCountDepth}
\title{Running SCnorm  checkCountDepth test.}
\usage{
checkCountDepth(group = c("sudo", "docker"), data.folder = getwd(),
  counts.matrix, conditions = NULL, FilterCellProportion = 0.1,
  FilterExpression = 0, ditherCounts = FALSE, outputName, nCores = 8)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs.}

\item{data.folder, }{a character string indicating the folder where output will be saved.}

\item{counts.matrix, }{a character string indicating the the name of tab delimited file  of cells un-normalized expression counts.}

\item{conditions, }{vector of condition labels, this should correspond to the columns of the un-normalized expression matrix. If not provided data is assumed to come from same condition/batch.}

\item{FilterCellProportion, }{a value indicating the proportion of non-zero expression estimates required to include the genes into the evaluation. Default is .10, and will not go below a proportion which uses less than 10 total cells/samples}

\item{FilterExpression, }{a value indicating exclude genes having median of non-zero expression below this threshold from count-depth plots}

\item{ditherCounts, }{whether to dither/jitter the counts, may be used for data with many ties, default is FALSE}

\item{outputName, }{specify the path and/or name of output files.}

\item{nCores, }{number of cores to use, default is detectCores() - 1.}
}
\value{
pdf with the cells counts distributions
}
\description{
This function executes a check on the data count-depth relationship used by SCnorm.
}
\examples{
\dontrun{
    #C1 data whole transcript
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    conditions=rep(1,288)
    checkCountDepth(group="docker", data.folder=getwd(),
    counts.matrix="singlecells_counts.txt", conditions=conditions,
    FilterCellProportion=0.1, FilterExpression=0, ditherCounts=FALSE,
    outputName="singlecells_counts", nCores=8)

    #an other example with UMI 3' end analysis
    system("wget http://130.192.119.59/public/example_UMI.txt.zip")
    unzip("example_UMI.txt.zip")
    conditions=rep(1,12)
    checkCountDepth(group="docker", data.folder=getwd(), counts.matrix="example_UMI.txt",
    conditions=conditions, FilterCellProportion=0.1, FilterExpression=0,
    ditherCounts=TRUE, outputName="example_UMI", nCores=8)

}
}
