% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterIdentification.R
\name{clusterIdentification}
\alias{clusterIdentification}
\title{Cluster identification}
\usage{
clusterIdentification(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nCluster,
  separator
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{which nCluster ins interesting to use for this analysis}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
Csv file with correct cluster name
}
\description{
This function executes a ubuntu docker that normalize all the permutation finding the rispective number of cluster respect the main clustering, the P0
}
\examples{
\dontrun{
 clusterIdentification(group,scratch.folder,file,nCluster,separator)#
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
