% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterNgriph.R
\name{clusterNgriph}
\alias{clusterNgriph}
\title{Defining with griph the range of  number of clusters to be used with SIMLR}
\usage{
clusterNgriph(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nPerm,
  permAtTime,
  percent,
  separator,
  logTen = 0,
  seed = 111
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to perform}

\item{permAtTime, }{number of permutations that can be computed in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{seed, }{important value to reproduce the same results with same input}
}
\value{
A folder Results containing a folder with the name of the experiment, which contains: VioPlot of silhouette cells value for each number of cluster used, a folder with the number of clusters selected as optimal by griph, which contains: clusterP file with clustering results for each permutation, killedCell file with removed cells in each permutation, clustering.output a sommarize file with general information for each cells, hist.pdf describing the number of times a specific number of clusters was used during permutations
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate the range of optimal number of clusters using griph algorithm. For more info see rCASC vignette.
}
\examples{
\dontrun{
 clusterNgriph(group="docker",scratch.folder="/data/scratch/",file=paste(getwd(), "bmsnkn_5x100cells.txt", sep="/"), nPerm=16, permAtTime=8, percent=10, separator="\t",logTen=0, seed=111)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
