% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustersFeatures.R
\name{clustersFeatures}
\alias{clustersFeatures}
\title{A function to extract priority lists of genes from ANOVAlike output}
\usage{
clustersFeatures(group = c("sudo", "docker"), data.folder, logFC.table,
  counts.table, delta = 0.5)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{logFC.table, }{a character string indicating the logFC_filtered_DE_ file generated by anovaLike function.}

\item{counts.table, }{a character string indicating the counts table file.}

\item{delta, }{in the minimal distance between the max value of FC for a gene in a cluster of interest and the nearest other max FC in any of the other clusters. This value define the minimal distance with respect to the same gene in an other cluster to identify it as a cluster specific gene.}
}
\value{
Three tab delimited files file with prefix onlyUP\_, followed by the counts table name, i.e. the count table only containing the selected genes. onlyUP\_, followed by logFC_filtered_DE_,  the table containing logFC only for the selected genes, onlyUP_clusters_specific_genes.txt, which contains the list of specific genes associated with the corresponding cluster.
}
\description{
This function extracts clusters specific gene lists from ANOVAlike output. IMPORTANT: This analysis works only on up modulated genes, i.e. genes more expressed in the clusters under analysis with respect to the reference cluster
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/clusters.features.zip")
unzip("clusters.features.zip")
setwd("clusters.features")

    clustersFeatures(group="docker", data.folder=getwd(),
        logFC.table="logFC_filtered_DE_annotated_setPace_10000_noC5_reordered.txt",
        counts.table="annotated_setPace_10000_noC5_reordered.txt"
        cluster.file="annotated_setPace_10000_noC5_clustering.output.txt", delta=0.5)
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino
}
