% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossLabel.R
\name{crossLabel}
\alias{crossLabel}
\title{Cross Label}
\usage{
crossLabel(group = c("sudo", "docker"), scratch.folder, file, nCluster,
  separator, finalName)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{format, }{count matrix format "csv", "txt"..}

\item{fileName, }{name of the result label plot}
}
\value{
Csv file with cluster component informations and dataPlot and pie chart based on true Label
}
\description{
This function executes a ubuntu docker that produces informations about the true clusters
}
\examples{
\dontrun{
 crossLabel("docker","/home/lucastormreig/CASC5.0/7_crossLabel/scratch/","/home/lucastormreig/CASC5.0/7_crossLabel/Data/New_Buettner",3,",")#
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
