% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterZeros.R
\name{filterZeros}
\alias{filterZeros}
\title{Plotting the distribution of zeros in cells eliminating all genes without counts}
\usage{
filterZeros(file, threshold = 0, sep)
}
\arguments{
\item{file, }{a character string indicating the path of the file tab delimited  of cells un-normalized expression counts}

\item{threshold, }{a number from 0 to 1 indicating the fraction of max accepted zeros in each gene. 0 is set as default and it eliminates only genes having no expression in any cell.}

\item{sep, }{separator used in count file, e.g. '\\t', ','}
}
\value{
a PDF providing zeros distributions before removal of all genes without counts. A tab delimited file with the prefix *filtered* in which the filtered data are saved.
}
\description{
This function plots the zeros distributions in cells and removes genes without counts
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    filterZeros(file=paste(getwd(),"singlecells_counts.txt",sep="/"), threshold=0.1, sep="\\t")
}
}
