% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesUmi.R
\name{genesUmi}
\alias{genesUmi}
\title{Plotting genes to UMIs relationship.}
\usage{
genesUmi(file, umiXgene = 3, sep)
}
\arguments{
\item{file, }{a character string indicating the path of the file tab delimited  of cells un-normalized expression counts.}

\item{umiXgene, }{a integer defining how many UMI are required to call a gene as present. default: 3}

\item{sep, }{separator used in count file, e.g. '\\t', ','}
}
\value{
pdf with the cells counts distributions: genes.umi.pdf
}
\description{
This function create a plot in which genes are plotted with respect to total counts UMI for each cell.
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/testSCumi_mm10.csv.zip")
unzip("testSCumi_mm10.csv.zip")
genesUmi(file=paste(getwd(),"testSCumi_mm10.csv",sep="/"), umiXgene=3, sep=",")
}
}
