% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5tocsv.R
\name{h5tocvs}
\alias{h5tocvs}
\title{h5 to csv}
\usage{
h5tocvs(group = c("sudo", "docker"), file)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{file, }{path of the h5 file. Full path MUST be included.}
}
\value{
a csv file results_cellranger
}
\description{
This function takes h5 file from cellranger output and convert it in csv table.
}
\examples{
\dontrun{
home <- getwd()
library(rCASC)
#download from https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE106268 the GSE106268_RAW.tar file
system("tar xvf GSE106268_RAW.tar")

h5tocvs(group="docker", file=paste(getwd(),"GSM2833284_Naive_WT_Rep1.h5",sep="/"))
}


}
\author{
Raffaele Calogero, raffaele [dot] calogero [at] unito [dot] com, University of Torino
}
