% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfc.R
\name{hfc}
\alias{hfc}
\title{hfc, heatmap for clustering}
\usage{
hfc(group = c("sudo", "docker"), scratch.folder, file, nCluster, separator,
  lfn, geneNameControl = 1, status)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{nCluster, }{nCluster interested for the analysis}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{lfn, }{name of the list of genes}

\item{geneNameControl, }{0 if the matrix has gene name without ENSEMBL code.1 if the gene names is formatted like this : ENSMUSG00000000001:Gnai3. If the gene names is only ensamble name you have to run SCannoByGtf before start using this script.}

\item{status, }{0 if is raw count, 1 otherwise}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{format, }{matrix count format, "csv", "txt"}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function generate heatmap and other plot based on clustering and on a specific gene list
}
\examples{
\dontrun{
hfc("docker","/home/lucastormreig/scratch/","/home/lucastormreig/CASC7.2/6_1hfc/Data/random_10000_filtered_annotated_lorenz_naive_penta2_0",6,",","naive")#
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
