% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseKeep.R
\name{houseKeep}
\alias{houseKeep}
\title{houseKeep Filtering}
\usage{
houseKeep(group = c("sudo", "docker"), scratch.folder, file, separator,
  logTen, geneNameControl = 1, houseK, topCell)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{0 if is raw count, 1 otherwise}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}
}
\value{
return matrix filtered by the best cells that have housekeepin genes most expressed
}
\description{
This function filter your matrix count on a specific gene house keepin list
}
\examples{
\dontrun{
housekeep("docker","/home/lucastormreig/scratch/","/home/lucastormreig/CASC7.2/6_1hfc/Data/random_10000_filtered_annotated_lorenz_naive_penta2_0",6,",","naive")#
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
