% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrationCircos.R
\name{integrationCircos}
\alias{integrationCircos}
\title{integrationCircos}
\usage{
integrationCircos(
  group = c("sudo", "docker"),
  scratch.folder,
  gsea.file,
  isc.file,
  XYpb.file,
  pblkae.file,
  outputFolder
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{gsea.file, }{a character string indicating the path of the final_score.csv  generated with gseaXLmHG, file, with file name and extension included.}

\item{isc.file, }{a character string indicating the path of the final_score.csv  generated with seuratIntegrationPermutation, file, with file name and extension included. included.}

\item{XYpb.file, }{a character string indicating the path of the final_score.csv  generated with toprnk, file, with file name and extension included.}

\item{pblkae.file, }{a character string indicating the path of the final_score.csv  generated with integrationPblkae, file, with file name and extension included.}

\item{outputFolder, }{where results are placed}
}
\value{
A picture called integrated_score.png and a file called integrated_score.csv and all the final_scores.csv used to produce the integrated results.
}
\description{
This function execute toprnk analysis which search for correspondence between clusters of two different experiments using clusters-pseudobulks, zscored on rows, and the cluster specific genes from comet analysis. Thus, the function clustersBulk and cometsc have to be run in the two datasets before their comparison.
}
\examples{
\dontrun{
library(rCASC)
integrationCircos(group="docker", 
                  scratch.folder="/scratch", 
                  gsea.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/GSEA/final_score.csv",
                  isc.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/ISC/final_score.csv",
                  XYpb.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/XYpb/XYpb_final_score.csv",
                  pblkae.file="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2/pblkAE/final_score.csv", 
                  outputFolder="/data/reanalysis_on_AIsc/comparing_CRC0327/NT1_NT2"
)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
