% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrationPblkae.R
\name{integrationPblkae}
\alias{integrationPblkae}
\title{integrationPblkae}
\usage{
integrationPblkae(
  group = c("sudo", "docker"),
  scratch.folder,
  fileX = NULL,
  fileY = NULL,
  outputFolder,
  type = c("inter", "intra"),
  stats = c("anovalike", "pairwise")
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{fileX, }{a character string indicating the path to the total.csv.for the 1st dataset to be integrated. Total.csv is generated with autoencoder4pseudoBulk. File, with file name and extension included.}

\item{fileY, }{a character string indicating the path to the total.csv.for the 2nd dataset to be integrated. Total.csv is generated with autoencoder4pseudoBulk. File, with file name and extension included.}

\item{outputFolder, }{where results are placed}

\item{type, }{two values inter, intra. Inter refers to comparison among clusters of two independent experiments. Intra comparisons among clusters of the same experiment}

\item{stats, }{two values anovalike, pairwise. Anovalike refers to comparison among clusters with respect to a pseudo reference sample. pairwise comparisons among all clusters by pairwise comparison}
}
\value{
a folder called psblkAE, which contains file called final_score.csv and all the intermediate files used to produce the integrated results.
}
\description{
This function execute integrationPblkae analysis which search for correspondence between clusters of two different experiments using clusters-pseudobulks generated using sparsely connected autoencoders. Thus, the function autoencoder4pseudoBulk has to be run in the two datasets before their comparison.
}
\examples{
\dontrun{
 #inter
 library(rCASC)
 integrationPblkae(group="docker", 
        scratch.folder="/scratch", 
        fileX="/data/clusters_association_paper/setA1_set1/setA1/VandE/Results/setA1/permutation/total.csv",
        fileY="/data/clusters_association_paper/setA1_set1/set1/VandE/Results/set1/permutation/total.csv",
        outputFolder="/data/clusters_association_paper/setA1_set1",
        type="inter",
        stats="anovalike"
 )
 
 #intra
 library(rCASC)
 integrationPblkae(group="docker", 
        scratch.folder="/scratch", 
        fileX="/data/clusters_association_paper/setA1_set1/setA1/VandE/Results/setA1/permutation/total.csv",
        outputFolder="/data/clusters_association_paper/setA1_set1/setA1",
        type="intra",
        stats="pairwise"
 )
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
