% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenzFilter.R
\name{lorenzFilter}
\alias{lorenzFilter}
\title{A function to handle sigle cell Lorenz Quality filter for Single-cells}
\usage{
lorenzFilter(group = c("sudo", "docker"), scratch.folder, data.folder,
  matrixName, p_value, umiXgene = 3, format, separator)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{p_value, }{threshold to be used for the filtering}

\item{umiXgene, }{a integer defining how many UMI are required to call a gene as present. default: 3}

\item{format, }{counts file extension, "txt", "csv"}

\item{separator, }{separator used in count file, e.g. '\\t', ','}
}
\value{
output will be in the same format and with the same separator of input. It also returns a PDF with the genes vs UMI plot, discarded_cells.pdf. In blue discarded cells are shown.
}
\description{
This function executes a docker that produces ....
}
\examples{
\dontrun{
 system("wget http://130.192.119.59/public/testSCumi_mm10.csv.zip")
 unzip("testSCumi_mm10.csv.zip")
 library("CASC")
 lorenzFilter(group="docker", scratch.folder="/data/scratch",
          data.folder=getwd(), matrixName="testSCumi_mm10", p_value=0.05, umiXgene=3, format="csv", separator=',')
}

}
\author{
Name Family name, myemail [at] somewhere [dot] org, Affiliation
}
