% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMatrix.R
\name{mergeMatrix}
\alias{mergeMatrix}
\title{Merge matrix}
\usage{
mergeMatrix(
  group = c("sudo", "docker"),
  scratch.folder,
  file1,
  file2,
  separator1,
  separator2,
  name1 = "NULL",
  name2 = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file1, }{a character string indicating the path of the first matrix to be merged, with file name and extension included}

\item{file2, }{a character string indicating the path of the second matrix to be merged, with file name and extension included}

\item{separator1, }{separator used in count file, e.g. '\\t', ','}

\item{separator2, }{separator used in count file, e.g. '\\t', ','}

\item{name1, }{the name that will be the prefix in the header of the matrix1 cells , null will not add any prefix}

\item{name2, }{the name that will be the prefix in the header of the matrix2 cells , null will not add any prefix}
}
\value{
a merged matrix
}
\description{
This function executes a ubuntu docker that merge two matrix
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")
subSetCell(group="docker", scratch.folder="/data/scratch", 
           file=paste(getwd(), "annotated_setPace_10000_noC5.txt",sep="/"), 
           separator="\t", cells.number=200)
mergeMatrix(group="docker", scratch.folder="/data/scratch", 
            file1=paste(getwd(),"annotated_setPace_10000_noC5.txt", sep="/"),
            file2=paste(getwd(),"subset_200_annotated_setPace_10000_noC5.txt", sep="/"), 
            separator1="\t",separator2="\t",name1="test1",name2="test2")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
