% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitoriboumi.R
\name{mitoRiboUmi}
\alias{mitoRiboUmi}
\title{MitoRiboUmi}
\usage{
mitoRiboUmi(group = c("sudo", "docker"), scratch.folder, file, separator,
  gtf.name, bio.type, umiXgene)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{gtf.name, }{name for the gtf file tu be used.}

\item{bio.type, }{ENSEMBL biotype of interest}

\item{umiXgene, }{umi to filter genes.}
}
\value{
Csv file with cluster component informations and dataPlot and pie chart based on true Label
}
\description{
This function executes a ubuntu docker that produces informations ribosomal mitochondrial fraction in each cell
}
\examples{
\dontrun{
   system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
   system("gzip -d singlecells_counts.txt.gz")
   mitoRiboUmi(group="docker",scratch.folder="/data/scratch/",file=paste(getwd(), "singlecells_counts.txt", sep="/"),separator="\\t", gtf="genome.gtf", biotype="protein_coding")
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
