% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permAnalysis.R
\name{permAnalysis}
\alias{permAnalysis}
\title{Permutation Analysis}
\usage{
permAnalysis(group = c("sudo", "docker"), scratch.folder, file, range1,
  range2, separator, sp, clusterPermErr = 0.05, maxDeltaConfidence = 0.01,
  minLogMean = 0.05)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{range1, }{First number of cluster that has to be analyzed}

\item{range2, }{Last number of cluster that has to be analyzed}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{sp, }{minimun number of percentage of cells that has to be in common between two permutation to be the same cluster.}

\item{clusterPermErr, }{error that can be done by each permutation in cluster number depicting.Default = 0.05}

\item{maxDeltaConfidence, }{max value for Delta confidence for genes feature selection}

\item{minLogMean, }{min value for Log mean for genes feature selection}
}
\value{
stability plot for each nCluster,two files with score information for each cell for each permutation.
}
\description{
This function analyze the data that came up from permutationClustering script.
}
\examples{
\dontrun{
permAnalysis("docker","path/to/scratch","path/to/data/TOTAL",3,4,",",0.8)#
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
