% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationMovie.R
\name{permutationMovie}
\alias{permutationMovie}
\title{Permutation Movie}
\usage{
permutationMovie(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nCluster,
  separator,
  framePP,
  permutationNumber
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nCluster, }{cluster folder generated by simlBootstrap or tsneBootstrap to be analysed}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{framePP, }{Number of frame for each permutation}

\item{permutationNumber, }{Number of random permutation,have to be less or the same then the total permutation}
}
\value{
Csv file with correct cluster name
}
\description{
This function executes a ubuntu docker that create a video showing the cluster changing of all the cells
}
\examples{
\dontrun{
 permutationMovie("docker","/home/lucastormreig/CASC5.0/5_permutationMovie/scratch/",file,7,",",200,5)#
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
