% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCannoByGtf.R
\name{scannobyGtf}
\alias{scannobyGtf}
\title{Annotating single cell counts table using ENSEMBL gtf and refGenome CRAN package}
\usage{
scannobyGtf(group = c("docker", "sudo"), file, gtf.name,
  biotype = NULL, mt = c(TRUE, FALSE), ribo.proteins = c(TRUE,
  FALSE), umiXgene = 3)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{file, }{a character string indicating the folder where input data are located and where output will be written and matrix name "/bin/users/matrix.csv". The system recognize automatically csv as comma separated files and txt as tab separated file}

\item{gtf.name, }{a character string indicating the ENSEMBL gtf file}

\item{biotype, }{a character string the biotypes of interest}

\item{mt, }{a boolean to define if mitocondrial genes have to be removed, FALSE mean that mt genes are removed}

\item{ribo.proteins, }{a boolean to define if ribosomal proteins have to be removed, FALSE mean that ribosomal proteins (gene names starting with rpl or rps) are removed}

\item{umiXgene, }{a integer defining how many UMI are required to call a gene as present. default: 3}
}
\value{
one file: annotated_counts table, where ensembl ids are linked to gene symbols and a PDF showing the effect of ribo and mito genes removal.
}
\description{
This function executes the docker container annotate.1, where refGenome is used to annotate a single cell counts table with ensembl gene ids on first column using ENSEMBL GTF annotation
}
\examples{
\dontrun{
        system("wget http://130.192.119.59/public/testSCumi_mm10.csv.zip")
     library(casc)
     system("unzip testSCumi_mm10.csv.zip")
     #filtering low quality cells
     lorenzFilter(group="docker",scratch.folder="/data/scratch/",
                  file=paste(getwd(),"testSCumi_mm10.csv",sep="/"),
                  p_value=0.05,separator=',')
     #running annotation and removal of mit and ribo proteins genes
     #download mouse GTF for mm10
     system("wget ftp://ftp.ensembl.org/pub/release-92/gtf/mus_musculus/Mus_musculus.GRCm38.92.gtf.gz")
     system("gunzip Mus_musculus.GRCm38.92.gtf.gz")
     scannobyGtf(group="docker", file=paste(getwd(),"lorenz_testSCumi_mm10.csv",sep="/"),
                   gtf.name="Mus_musculus.GRCm38.92.gtf",
biotype="protein_coding", mt=TRUE, ribo.proteins=TRUE,umiXgene=3)
}

}
\author{
Raffaele Calogero
}
