% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratprior.R
\name{seuratPrior}
\alias{seuratPrior}
\title{seuratprior}
\usage{
seuratPrior(group = c("sudo", "docker"), scratch.folder, file, separator,
  logTen = 0, seed = 1111, PCADim, geneNumber, nCluster)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{seed, }{important value to reproduce the same results with same input}

\item{PCADim, }{dimensions of PCA for seurat clustering}

\item{geneNumber, }{numbers of specific genes for each clusters}

\item{nCluster, }{number of cluster analysis.}
}
\value{
Plot with PCA scores is provided to detect the PCA dimensions to be used in Seurat clustering algorithm
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering.
}
\examples{
\dontrun{
 permutationClustering("docker","/home/lucastormreig/CASC2.0/permutationClustering/scratch/","/home/lucastormreig/CASC2.0/permutationClustering/Data/TOTAL.csv",4,2,10,3,4,separator=",",logTen=0,clustering="SIMLR",perplexity=0)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
