% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STLearnPermutation.R
\name{STLearnPermutation}
\alias{STLearnPermutation}
\title{StLearn Permutation}
\usage{
STLearnPermutation(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  nPerm,
  permAtTime,
  percent,
  separator,
  logTen = 0,
  pcaDimensions = 5,
  seed = 1111,
  sparse = FALSE,
  format = "NULL"
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of randomly selected cells removed in each permutation}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{pcaDimensions, }{0 for automatic selection of PC elbow.}

\item{seed, }{important value to reproduce the same results with same input}

\item{sparse, }{boolean for sparse matrix}

\item{format, }{output file format csv or txt}
}
\value{
To write
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering 
 using seurat (updated to 2020) and a stlearn normalized expression matrix as input.
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino

Giovanni Motterle, giovanni [dot] motterle [at] studenti [dot] univr [dot] it, University of Verona
}
