% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSetCell.R
\name{subSetCell}
\alias{subSetCell}
\title{Subset Cell}
\usage{
subSetCell(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  cells.number
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with file name and extension included}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{cells.number, }{number of cells to be extracted}
}
\value{
subsetted matrix
}
\description{
This function executes a ubuntu docker that subset a matrix
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/annotated_setPace_10000_noC5.txt.zip")
unzip("annotated_setPace_10000_noC5.txt.zip")

subSetCell<- function(group="docker", scratch.folder="/data/scratch", 
                       file=paste(getwd(), "annotated_setPace_10000_noC5.txt",sep="/"), 
                       separator="\t", cells.number=200)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
