% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topX.R
\name{topx}
\alias{topx}
\title{A function to selectec top X on the basis of gene/transcript expression}
\usage{
topx(group = c("sudo", "docker"), file, threshold, separator,
  logged = FALSE, type = c("expression", "variance"))
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{file, }{a character string indicating the path of the file. IMPORTANT: full path to the file MUST be included}

\item{threshold, }{integer used for filtering indicate the number of top expressed/vaying genes to be selected}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logged, }{boolean TRUE or FALSE, if FALSE gene expression data are log10 transformed before being plotted.}

\item{type, }{expression refers to the selection of the top expressed genes, variance to the the selection of the top variable genes}
}
\value{
a filtered tab delimited file and a histogram of the gene by gene total expression
}
\description{
This function select the X top genes given a user defined threshold
}
\examples{
\dontrun{

 system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
 system("gzip -d singlecells_counts.txt.gz")
 topx(group="docker", file=paste(getwd(), "singlecells_counts.txt", sep="/"),
       threshold=10000, logged=FALSE, type="expression", separator="\\t")
       
 topx(group="docker", file=paste(getwd(), "singlecells_counts.txt", sep="/"), 
       threshold=10000, logged=FALSE, type="variance", separator="\\t")
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, UNITO
}
