% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperMixModelsUmap.R
\name{wrapperMixModelsUmap}
\alias{wrapperMixModelsUmap}
\title{wrapperMixModelsUmap}
\usage{
wrapperMixModelsUmap(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  geneList,
  separator,
  k,
  seed,
  epochs,
  finalName
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the count matrix}

\item{geneList, }{a character string indicating the path of the geneList matrix (no header, no row names)}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{k, }{number of soft labels to divide the density}

\item{seed, }{number for reproducibility}

\item{epochs, }{epochs number for umap algorithm}

\item{finalName, }{name for output files}
}
\value{
plot
}
\description{
This function executes a ubuntu docker that performs wrapperMixModelsUmap
}
\examples{
\dontrun{
dir.create("scratch")
geneVisualization(group=c("sudo"), scratch.folder=paste(getwd(),"/scratch",sep=""), file=paste(getwd(),"/setA.csv",sep=""),geneList=paste(getwd(),"/geneList.csv",sep=""),separator=",",k=2)
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
